### 2025-01-29 #7736 @editorialbot
[PRE REVIEW]: Pop2net: A Python package for bipartite networks in agent-based simulations
**Submitting author:** @mariuzka (<a href="http://orcid.org/0000-0002-8619-3362">Marius Kaffai</a>)
**Repository:** https://github.com/mariuzka/pop2net
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.0
