### 2025-01-29 #7736 @mariuzka
> [@mariuzka](https://github.com/mariuzka) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

Hello @editorialbot ,
One potential reviewer could be Joël Foramitti (jofmi), the creator of AgentPy.
