+++ docs/api/magiclocation.rst
@@ -1,9 +1,0 @@
-.. currentmodule:: pop2net
-
-MagicLocation
-=============
-
-The MagicLocation class is an interface to design specific network structures using the Creator.
-
-.. autoclass:: MagicLocation
-    :members:
