### 2025-01-29 #7736 @danielskatz
👋 @mariuzka - thanks for your submission.

1. Can you please work on your README? As you can see in the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), there should be more content here, such as some explanation of the software, installations instructions, contribution info, etc.

2. Please suggest a few more reviewers, ideally around 5 total, perhaps using the JOSS list if it's helpful

3. Please work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
