+++ paper/paper.bib
@@ -4,10 +4,12 @@
 	volume = {24},
 	issn = {1460-7425},
 	shorttitle = {Social {Network} {Metric}-{Based} {Interventions}?},
+	doi = {10.18564/jasss.4571},
 	number = {3},
 	journal = {Journal of Artificial Societies and Social Simulation},
 	author = {Vermeulen, Ben and Müller, Matthias and Pyka, Andreas},
 	year = {2021},
+	keywords = {Agent-Based Model, Coronavirus, COVID-19, Epidemic, Policy Laboratory},
 	pages = {6},
 	file = {Social Network Metric-Based Interventions?:C\:\\Users\\ac135963\\Zotero\\storage\\YNREPJ6M\\6.html:text/html;Vermeulen et al_2021_Social Network Metric-Based Interventions.pdf:C\:\\Users\\ac135963\\Zotero\\storage\\5VUBUQ7A\\Vermeulen et al_2021_Social Network Metric-Based Interventions.pdf:application/pdf},
 }
@@ -34,28 +36,9 @@
 	month = oct,
 	year = {2021},
 	note = {Publisher: Public Library of Science},
-	keywords = {Agent-based modeling, COVID 19, Economic agents, Germany, Pandemics, Schools, Simulation and modeling, Surveys},
+	keywords = {Germany, Surveys, COVID 19, Agent-based modeling, Pandemics, Schools, Simulation and modeling, Economic agents},
 	pages = {e0259108},
 	file = {Full Text PDF:C\:\\Users\\ac135963\\Zotero\\storage\\R82AL6AY\\Kaffai und Heiberger - 2021 - Modeling non-pharmaceutical interventions in the C.pdf:application/pdf;Snapshot:C\:\\Users\\ac135963\\Zotero\\storage\\WFSBQWMB\\article.html:text/html},
-}
-
-@article{cortez_dynamics_2015,
-	series = {Complex {Adaptive} {Systems} {San} {Jose}, {CA} {November} 2-4, 2015},
-	title = {Dynamics of the {Schelling} {Social} {Segregation} {Model} in {Networks}},
-	volume = {61},
-	issn = {1877-0509},
-	url = {https://www.sciencedirect.com/science/article/pii/S1877050915029786},
-	doi = {10.1016/j.procs.2015.09.148},
-	abstract = {Thomas C. Schelling introduced a simple mathematical model to address the phenomenon of social segregation as a consequence of a natural evolution of elementary local rules. This model became a good framework to understand how local rules, like the selection of a better neighborhood for life, can produce macro-behaviors, like segregation in a population with different kind of individuals. In this work, with the aid of the Schelling's model, we characterize the dynamical evolution of segregation in random networks; analyze the performance of some segregation indices and the dynamics of this model for various network degrees. Finally, we provide probabilistic arguments to explain quantitatively the behavior of the dynamic of Schelling's segregation model in networks.},
-	language = {en},
-	urldate = {2023-04-03},
-	journal = {Procedia Computer Science},
-	author = {Cortez, Vasco and Rica, Sergio},
-	month = jan,
-	year = {2015},
-	keywords = {dynamic models., index of segregation, Network, Schelling model, segregation},
-	pages = {60--65},
-	file = {ScienceDirect Full Text PDF:C\:\\Users\\ac135963\\Zotero\\storage\\3WFYDRV6\\Cortez und Rica - 2015 - Dynamics of the Schelling Social Segregation Model.pdf:application/pdf;ScienceDirect Snapshot:C\:\\Users\\ac135963\\Zotero\\storage\\NQEEKG4R\\S1877050915029786.html:text/html},
 }
 
 @article{jiang_method_2022,
@@ -192,6 +175,38 @@
[…]
