### 2025-01-29 #7736 @mariuzka
> 2\. Please suggest a few more reviewers, ideally around 5 total, perhaps using the JOSS list if it's helpful

👋 @danielskatz,

I would like to suggest the following potential reviewers:

- Joël Foramitti (https://github.com/jofmi)
- Ewout ter Hoeven (https://github.com/EwoutH)
- Brennan Klein (https://github.com/jkbren)
- Jonas Pleyer (https://github.com/jonaspleyer)
- Shuang Song (https://github.com/SongshGeo)
- Janik Muires (https://github.com/janikmu)
