### 2025-01-31 33eaaf0
Update some docstrings
