+++ README.md
@@ -3,7 +3,21 @@
 
 # Pop2net
 
-Pop2net is a Python package designed to generate and manage relations between agents in agent-based models in a simple and flexible way.
+Pop2net is a Python package that combines many steps related to network generation and management for agent-based modeling using a bipartite approach: 
+In Pop2net, relations are represented as a graph of agents and locations. Locations act as a contact layer, representing specific contexts through which agents connect. 
+This makes it easy to generate and manage custom network structures in an agent-based model.
+Pop2net extends the existing agent-based modeling framework [AgentPy](https://github.com/jofmi/agentpy), while also providing functionality for generating networks compatible with other modeling frameworks.
 
-# Documentation
+
+## Installation
+
+You can install Pop2net using pip:
+
+```
+pip install pop2net
+```
+
+
+## Documentation
+
 Detailed documentation of all features and a collection of example analyses can be found [here](https://mariuzka.github.io/pop2net/).
