### 2025-03-05 #7736 @hugoledoux
@EwoutH Hello Ewout (weird that our paths cross again this way!) would you be interested in reviewing this for JOSS? I think you fit really well for it!
