### 2025-03-05 #7736 @hugoledoux
@jofmi you are an obvious candidate for a reviewer here, I hope you accept! Let me know.

oh I see you're not in our database of reviewers... Well, this is for a review for JOSS, the Journal of Open Source Software; see https://joss.theoj.org/.

OSS is a free, open-source, community-driven and developer-friendly online journal
(no publisher is seeking to raise revenue from the volunteer labour of researchers!).

The review process at JOSS is unique: it takes place in a GitHub issue, is open,
and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository
and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments
and improve their submission until acceptance (or withdrawal, if they feel unable to
satisfy the review).
