### 2025-03-17 0d298d3
fix LocationDesigner.n_agents bug
