### 2025-03-18 2f2baf7
remove developement installation
