+++ docs/installation.rst
@@ -11,23 +11,3 @@ run the following command on your console:
 .. code-block:: console
 
 	$ pip install pop2net
-
-
-
-Development
------------
-
-The most recent version of Pop2net can be cloned from Github:
-
-.. code-block:: console
-
-	$ git clone https://github.com/mariuzka/pop2net.git
-
-Once you have a copy of the source, you can install it using [Poetry](https://python-poetry.org/docs/):
-
-.. code-block:: console
-
-    $ poetry install
-
-
-.. _Github repository: https://github.com/mariuzka/pop2net
