### 2025-03-21 #7897 @EwoutH
@mariuzka @luerhard I'm still digging in, but I have some initial suggestions on the paper itself:

1. **Add the workflow visualization diagram** - The [about page](https://mariuzka.github.io/pop2net/) in the documentation includes a helpful workflow diagram showing the relationship between micro-level data, agent classes, and location instances. Including this visualization (or a simplified version) would give readers a clearer understanding of Pop2net's approach.

2. **Incorporate practical code examples** - Your paper would benefit from showing how the concepts translate into straightforward code:
   ```python
   # Creating locations and agents
   agent = p2n.Agent(model)
   location = p2n.Location(model)

   # Connecting them
   location.add_agent(agent)

   # Finding neighbors through shared locations
   agent.neighbors(location_labels=["School"])
   ```
   These examples would demonstrate the simplicity of the API and make the bipartite approach more concrete. Here's [an example](https://github.com/openjournals/joss-papers/blob/joss.07668/joss.07668/10.21105.joss.07668.pdf) how we did this in our Mesa paper (we might have done a little much).

3. **Highlight location types as organizational tools** - The paper mentions location types but could better explain how they effectively organize different relationship contexts. An example or visual might also help here:
   ```python
   # Custom location types for different social contexts
   class School(p2n.Location):
       pass

   class Workplace(p2n.Location):
       pass

   # Filtering neighbors by context
   agent.neighbors(location_labels=["School"])  # Only school connections
   ```
   This pattern enables modelers to implement context-specific behaviors, a notable advantage of your approach.

4. **Clarify distinction from population synthesizers** - The documentation states that "Pop2net is not a population synthesizer." Adding this distinction to your paper would help readers better understand your tool's purpose and scope.

5. **Expand on NetworkInspector functionality** - The paper briefly mentions "inspection tools," but could benefit from highlighting the visualization capabilities (`inspector.plot_bipartite_network()` / `inspector.plot_agent_network()`) and maybe include a few figures as examples.

6. **Better explain relationship with AgentPy** - The documentation describes Pop2net as "a plugin for AgentPy" while the paper mentions "extending" it. A clearer explanation of how a ABM / AgentPy developer might use Pop2net would be useful for your target audience.

7. Extending that, maybe also discuss why AgentPy specifically was extended in the first place, and if it could also work with other Python ABM libraries (like Mesa for example).

8. **Showcase network management features** - The paper could demonstrate the intuitive API for common operations:
   ```python
   # Network management for connecting multiple agents
   model.connect_agents(agents=model.agents, location_cls=School)

   # Finding shared contexts between agents
   agent.shared_locations(other_agent, location_labels=["School"])
   ```
   These methods significantly reduce the code typically needed for network manipulation.

9. **Illustrate bipartite-to-agent network projection** - The documentation shows how Pop2net handles the projection from bipartite to agent networks. This feature deserves mention as it simplifies network analysis.

I really like the cleanness of the API and the tutorial is written well.

Personally, I would be curious if we could make pop2net compatible with Mesa, but that's probably a discussion for another time.

To summarize the feedback, maybe the pop2net tutorial itself already states it:

> Often the best thing you can do to understand things is to look at them.

Hope this is useful!
