### 2025-03-21 mariuzka/pop2net#153 @EwoutH
Feedback on intro tutorial
A few points on the intro tutorial (in general I really like it!)

> ![Image](https://github.com/user-attachments/assets/a9e0b2f4-abe6-4b5a-a960-154d271b1f1a)

1. State which dot represents what. Red is the agent right?

> Just because we can, let's create some more edges by adding one agent to all locations:
> ```Python
> agent = model.agents[0]
> for location in model.locations:
>     location.add_agent(agent)
> ```

2. I seems it would be logical to add a function that adds a list/set of locations to an agent, and a function to add a list/set of agents to a location.

Maybe just allow functions like `add_location()` `add_agent()`, `remove_agent() ` to also take an iterable, then you don't need new function at all!

> The methods `agent.connect()` and `model.connect_agents()` simplify that process by creating an instance from a given location class and directly assigning the given agents to this location instance.
In the following, we connect `agent` with the second agent of the model `model.agents[1]` using `agent.connect()`:

3. It would be useful to explain which agents are connected with which locations (how that works under the hood).
