### 2025-03-21 #7897 @EwoutH
Posted some notes on the tutorial here:
- https://github.com/mariuzka/pop2net/issues/153
