### 2025-03-24 mariuzka/pop2net#153 @mariuzka
Hi,
Thanks for the great comments!

1. You're absolutely right—I’ll add a description explaining what each dot represents.
2. Great suggestion! The truth is, there are already functions for that; I just forgot to cover them in the tutorial. The functions are [agent.add_locations()](https://github.com/mariuzka/pop2net/blob/ffa1885c3a44188ad3e1dc0df753f970301ea6c9/src/pop2net/agent.py#L78) and [location.add_agents()](https://github.com/mariuzka/pop2net/blob/ffa1885c3a44188ad3e1dc0df753f970301ea6c9/src/pop2net/location.py#L48). I am going to add them to the tutorial.
3. Okay, I’ll add an explanation to the tutorial.
