### 2025-03-24 mariuzka/pop2net#153 @mariuzka
- [ ] Add description of dots in network plot
- [ ] Cover `location.add_agents()` etc. in tutorial
- [ ] Explain how `agent.connect()` etc. works
