### 2025-03-26 c221f04
set default value of location_cls in agent.connect()
