### 2025-03-26 4a935c7
auto-commit: fixes by ruff
