### 2025-03-26 a5afb60
upgrade to network viz to BokehGraph 1.0
