+++ pyproject.toml
@@ -40,7 +40,7 @@ agentpy = "^0.1.5"
 tabulate = "^0.9.0"
 seaborn = "^0.12.1"
 faker = "^15.1.1"
-bokehgraph = "^0.6.3"
+bokehgraph = "^1.0.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.2.0"
