### 2025-03-26 b026945
auto-commit: fixes by ruff
