### 2025-03-26 036ea9c
change edge_size and node_size
