### 2025-03-26 4a56521
adjust default values & update docstrings
