+++ src/pop2net/inspector.py
@@ -37,24 +37,25 @@ class NetworkInspector:
         edge_color: str = "black",
         edge_size: int = 1,
         node_size: int = 10,
+        node_alpha=0.5,
     ) -> None:
-        """Plots the two-mode network of agents and locations.
+        """Plots the bipartite network of agents and locations.
 
         Args:
-            agent_color (str, optional): The agent attribute that determines the
-                color of the agent nodes.
             agent_attrs (list | None, optional): A list of agent attributes that
                 should be shown as node attributes in the network graph. Defaults to None.
-            location_color (str, optional): The location attribute that determines the
-                color of the location nodes.
             location_attrs (list | None, optional): A list of location attributes that
                 should be shown as node attributes in the network graph. Defaults to None.
-            edge_alpha (str, optional): The node attribute that determines the edges' transparency.
+            agent_color (str | None, optional): The agent attribute that determines the
+            color of the agent nodes. Defaults to None.
+            location_color (str | None, optional): The location attribute that determines the
+            color of the location nodes. Defaults to None.
+            edge_alpha (str, optional): The edge attribute that determines the edges' transparency.
                 Defaults to "weight".
-            edge_color (str, optional): The node attribute that determines the edges' color.
-                Defaults to "black".
-            include_0_weights (bool, optional): Should edges with a weight of zero be included in
-                the plot? Defaults to True.
+            edge_color (str, optional): The color of the edges. Defaults to "black".
+            edge_size (int, optional): The size of the edges. Defaults to 1.
+            node_size (int, optional): The size of the nodes. Defaults to 10.
+            node_alpha (float, optional): The transparency of the nodes. Defaults to 0.5.
         """
         if agent_attrs is None:
             agent_attrs = ["type"]
@@ -103,6 +104,7 @@ class NetworkInspector:
             edge_color=edge_color,
             edge_size=edge_size,
             node_size=node_size,
+            node_alpha=node_alpha,
         )
 
     def plot_agent_network(
@@ -114,21 +116,23 @@ class NetworkInspector:
         include_0_weights: bool = True,
         edge_size: int = 1,
         node_size: int = 10,
+        node_alpha=0.5,
     ) -> None:
         """Plots the agent network.
 
         Args:
-            agent_color (str, optional): The node attribute that determines the
-                color of the nodes. If None, the node color represents whether
-                it is a location or an agent instance.
+            agent_color (str | None, optional): The agent attribute that determines the
+            color of the agent nodes. Defaults to None.
             agent_attrs (list | None, optional): A list of agent attributes that
                 should be shown as node attributes in the network graph. Defaults to None.
-            edge_alpha (str, optional): The node attribute that determines the edges' transparency.
+            edge_alpha (str, optional): The edge attribute that determines the edges' transparency.
                 Defaults to "weight".
-            edge_color (str, optional): The node attribute that determines the edges' color.
-                Defaults to "black".
+            edge_color (str, optional): The color of the edges. Defaults to "black".
             include_0_weights (bool, optional): Should edges with a weight of zero be included in
                 the plot? Defaults to True.
+            edge_size (int, optional): The size of the edges. Defaults to 1.
+            node_size (int, optional): The size of the nodes. Defaults to 10.
+            node_alpha (float, optional): The transparency of the nodes. Defaults to 0.5.
         """
         if agent_attrs is None:
             agent_attrs = ["type"]
@@ -154,6 +158,7 @@ class NetworkInspector:
             edge_alpha=edge_alpha,
             edge_color=edge_color,
             edge_size=edge_size,
+            node_alpha=node_alpha,
             
         )
 
@@ -168,24 +173,27 @@ class NetworkInspector:
         include_0_weights: bool = True,
[…]
