### 2025-03-27 dde2c4a
bump version to 0.4.0
