### 2025-03-27 47d612b
Inspector.eval_affiliations(): Remove dataframe metadata
