+++ src/pop2net/inspector.py
@@ -235,7 +235,7 @@ class NetworkInspector:
         )
 
         utils.print_header("Number of locations")
-        print(df1.location_label.value_counts())
+        print(df1.location_label.value_counts().to_frame())
 
         df1 = df1.groupby("location_label").describe()
         df1.columns = df1.columns.droplevel()
@@ -257,7 +257,7 @@ class NetworkInspector:
         df2 = df2.drop("count", axis=0)
 
         utils.print_header("Number of affiliated locations per agent")
-        print(df2)
+        print(df2.to_frame())
 
         if return_data:
             return df1, df2
@@ -390,6 +390,7 @@ class NetworkInspector:
 
             result_list.append(get_network_measures(nx_subgraph))
         return result_list
+    
 
     def location_crosstab(
         self,
