### 2025-03-27 1a0742e
auto-commit: fixes by ruff
