+++ _sources/Introduction/inspector.ipynb.txt
@@ -13,7 +13,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 15,
    "metadata": {},
    "outputs": [
     {
@@ -22,7 +22,7 @@
        "LocationList (6 objects)"
       ]
      },
-     "execution_count": 2,
+     "execution_count": 15,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -89,7 +89,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 16,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -105,14 +105,14 @@
     "The method `plot_bipartite_network()` plots the biparte network.\n",
     "Agents are displayed as circles and locations are displayed as squares.\n",
     "Using the arguments `agent_color` and `location_color`, agent and location attributes can be displayed as colors in the plot.\n",
-    "With the arguments `agent_attrs` and `location_attrs` you can add attributes that should be displayed for each node when you move the mouse pointer over it.\n",
+    "With the arguments `agent_attrs` and `location_attrs` you can add attributes that should be displayed in the tooltips for each node when you move the mouse pointer over it.\n",
     "\n",
     "In the following example I display the agent attribute `status` as color and add `id` to the list of shown attribute values per agent node."
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 17,
    "metadata": {},
    "outputs": [
     {
@@ -127,7 +127,7 @@
      "data": {
       "text/html": [
        "\n",
-       "  <div id=\"dae0c46e-fc9e-49b3-8a8e-da2b596a15d9\" data-root-id=\"p1085\" style=\"display: contents;\"></div>\n"
+       "  <div id=\"c4ddbb3b-2f7b-412d-bb56-8561b1dbfedc\" data-root-id=\"p2201\" style=\"display: contents;\"></div>\n"
       ]
      },
      "metadata": {},
@@ -135,12 +135,12 @@
     },
     {
      "data": {
[…]
