### 2025-04-23 #7897 @jofmi
Hi @mariuzka! Sorry that it took a while for me to complete the review. The package looks very good. It has a clear use case and provides a comprehensive set of tools for this approach. I was particularly impressed by the tutorial, which is very extensive and intuitive to read. Below you can find some minor comments for improvement. I agree with the points raised by @EwoutH, so I try not to repeat the same points here and focus only on additional thoughts.

Paper
- [ ] Please provide a short explanation in the beginning of what a bipartite network is. I think for a newcomer it would not immediately be clear. Maybe a graphic or example could be helpful.
- [ ] In the statement of need you say that other packages lack generalization while also being too abstract. These two critiques seem to contradict each other. Also, it does not seem that pop2net is about generalization, as it focuses on only one type of network. Maybe the statement of need could be improved by instead focusing on the advantages of pop2net's modelling approach (i.e. the three core features) in comparison to e.g. just using networkx graphs.
- [ ] "while also providing functionality for generating networks compatible with other modeling frameworks" - this sounds cool but i couldn't find an example how and with which other framework it can be used. As @EwoutH mentioned, it would be cool of course if you could do something like Model.agentpy() and Model.mesa() to use pop2net with different base packages, but that is probably a nice-to-have for the future. But I think you need to illustrate what is meant by this sentence.

Docs
- [ ] There's a warning in the lollipop example
- [ ] There's two markdown link errors in https://mariuzka.github.io/pop2net/Introduction/introduction_basics.html#Creating-edges-between-agents-and-locations (agent.add_locations and below )
- [ ] Typos (best to locate with search all files)
	- seperated > separated
	- follwoing > following
	- beteween > between
	- MeltLocationDesignerDesigner
