+++ src/pop2net/location.py
@@ -9,20 +9,18 @@ from . import agent as _agent
 from . import model as _model
 
 
-class Location(Object):
+class Location:
     """Base class for location objects."""
 
     label: str | None = None
 
-    def __init__(self, model: _model.Model) -> None:
+    def __init__(self) -> None:
         """Location constructor.
 
         Args:
             model: Model this Location should be associated with.
         """
-        super().__init__(model)
-        self.model = model
-        self.model.add_location(self)
+        self.env = None
         self.label = self.__class__.__name__ if self.label is None else self.label
 
     @property
@@ -32,7 +30,7 @@ class Location(Object):
         Returns:
             List of agents at this location.
         """
-        return self.model.agents_of_location(self)
+        return self.env.agents_of_location(self)
 
     def add_agent(self, agent: _agent.Agent, weight: float | None = None) -> None:
         """Assigns the given agent to this location.
@@ -43,7 +41,7 @@ class Location(Object):
                 Defaults to None. If weight is set to None, the weight is generated by
                 location.weight(), which returns 1 by default.
         """
-        self.model.add_agent_to_location(self, agent=agent, weight=weight)
+        self.env.add_agent_to_location(self, agent=agent, weight=weight)
 
     def add_agents(self, agents: list, weight: float | None = None) -> None:
         """Add multiple agents at once.
@@ -64,7 +62,7 @@ class Location(Object):
         Args:
             agent: Agent that is to be removed.
         """
-        self.model.remove_agent_from_location(
+        self.env.remove_agent_from_location(
             location=self,
             agent=agent,
         )
@@ -87,7 +85,7 @@ class Location(Object):
         Returns:
             AgentList: A list of all agents at this location except the passed agent.
         """
-        agents = self.model.agents_of_location(self)
+        agents = self.env.agents_of_location(self)
         agents.remove(agent)
         return agents
 
@@ -100,7 +98,7 @@ class Location(Object):
             agent (Agent): The agent.
             weight (float): The weight.
         """
-        self.model.set_weight(
+        self.env.set_weight(
             agent=agent,
             location=self,
             weight=weight,
@@ -115,7 +113,7 @@ class Location(Object):
         Returns:
             Edge weight.
         """
-        return self.model.get_weight(agent=agent, location=self)
+        return self.env.get_weight(agent=agent, location=self)
 
     def weight(self, agent) -> float:  # noqa: ARG002
         """Generates the edge weight between a given agent and the location instance.
