+++ src/pop2net/model.py
@@ -60,7 +60,7 @@ class Model:
         Returns:
             dict: A dictionary which stores the model's agents by their id.
         """
-        return {agent.id: agent for agent in self.agents}
+        return {agent.id_p2n: agent for agent in self.agents}
 
     @property
     def locations_by_id(self) -> dict:
@@ -69,7 +69,7 @@ class Model:
         Returns:
             dict: A dictionary which stores the model's locations by their id.
         """
-        return {location.id: location for location in self.locations}
+        return {location.id_p2n: location for location in self.locations}
 
     @property
     def locations(self) -> LocationList:
@@ -89,8 +89,8 @@ class Model:
         Args:
             agent: Agent to be added to the environment.
         """
-        if not self.g.has_node(agent.id):
-            self.g.add_node(agent.id, bipartite=0, _obj=agent)
+        if not self.g.has_node(agent.id_p2n):
+            self.g.add_node(agent.id_p2n, bipartite=0, _obj=agent)
 
     def add_agents(self, agents: list) -> None:
         """Add agents to the environment.
@@ -110,8 +110,8 @@ class Model:
         Args:
             location: Location to be added to the environment.
         """
-        if not self.g.has_node(location.id):
-            self.g.add_node(location.id, bipartite=1, _obj=location)
+        if not self.g.has_node(location.id_p2n):
+            self.g.add_node(location.id_p2n, bipartite=1, _obj=location)
 
     def add_locations(self, locations: list) -> None:
         """Add multiple locations to the environment at once.
@@ -145,14 +145,14 @@ class Model:
             Exception: Raised if the agent does not exist in the environment.
         """
         # TODO: Create custom exceptions
-        if not self.g.has_node(location.id):
+        if not self.g.has_node(location.id_p2n):
             msg = f"Location {location} does not exist in Environment!"
             raise Exception(msg)
-        if not self.g.has_node(agent.id):
+        if not self.g.has_node(agent.id_p2n):
             msg = f"Agent {agent} does not exist in Environment!"
             raise Exception(msg)
 
-        self.g.add_edge(agent.id, location.id, **kwargs)
+        self.g.add_edge(agent.id_p2n, location.id_p2n, **kwargs)
         self.set_weight(agent=agent, location=location, weight=weight)
 
     def remove_agent(self, agent: _agent.Agent) -> None:
@@ -163,8 +163,8 @@ class Model:
         Args:
             agent: Agent to be removed.
         """
-        if self.g.has_node(agent.id):
-            self.g.remove_node(agent.id)
+        if self.g.has_node(agent.id_p2n):
+            self.g.remove_node(agent.id_p2n)
 
     def remove_agents(self, agents: list) -> None:
         """Remove multiple agents from the environment at once.
@@ -183,8 +183,8 @@ class Model:
         Args:
             location: Location to be removed.
         """
-        if self.g.has_node(location.id):
-            self.g.remove_node(location.id)
+        if self.g.has_node(location.id_p2n):
+            self.g.remove_node(location.id_p2n)
 
     def remove_locations(self, locations: list) -> None:
         """Remove multiple locations at once.
@@ -211,15 +211,15 @@ class Model:
             Exception: Raised if the agent does not exist in the environment.
         """
         # TODO: use custom exceptions
-        if not self.g.has_node(location.id):
+        if not self.g.has_node(location.id_p2n):
             msg = f"Location {location} does not exist in Environment!"
             raise Exception(msg)
-        if not self.g.has_node(agent.id):
+        if not self.g.has_node(agent.id_p2n):
             msg = f"Agent {agent} does not exist in Environment!"
             raise Exception(msg)
 
[…]
