+++ src/pop2net/creator.py
@@ -2,12 +2,14 @@
 
 from __future__ import annotations
 
+import copy
 import itertools
 import math
 import random
 import warnings
 
 import pandas as pd
+
 
 import pop2net as p2n
 import pop2net.utils as utils
@@ -20,8 +22,9 @@ class Creator:
 
     def __init__(
         self,
-        model: p2n.Model,
+        env: p2n.Model,
         seed: int = None,
+        model = None #TODO: add type hints ap.Model and mesa.Model
     ) -> None:
         """Instantiate a creator for a specific model.
 
@@ -29,10 +32,11 @@ class Creator:
             model (p2n.Model): Model, for which a population should be created
             seed (int, optional): A seed for reproducibility. Defaults to 999.
         """
-        self.model = model
+        self.env = env
         self.seed = seed
         self.rng = random.Random(seed)
-        self._dummy_model = p2n.Model(parameters=self.model.p)
+        self.model = model
+        self._dummy_model = type(self.model)()
         self._temp_agent_attrs = ["_P2NTEMP_split_values", "_P2NTEMP_melt_location_weight"]
 
     def _create_dummy_location(self, designer) -> p2n.Location:
@@ -42,7 +46,8 @@ class Creator:
             (designer,) if lc is None else (designer, designer.location_class),
             {},  # TODO: warum funktioniert das hier nicht?: {"label": designer.label},
         )
-        location = cls(model=self._dummy_model)
+        #location = cls(model=self._dummy_model)
+        location = cls()
         location.label = (
             designer.label if designer.label is not None else utils._get_cls_as_str(designer)
         )
@@ -149,7 +154,7 @@ class Creator:
             A list of agents.
         """
         if clear:
-            self.model.remove_agents(self.model.agents)
+            self.env.remove_agents(self.env.agents)
 
         if df is not None:
             df = df.copy()
@@ -158,7 +163,10 @@ class Creator:
             agents = []
             for _, row in df.iterrows():
                 if agent_class_dict is None:
-                    agent = agent_class(model=self.model)
+                    if self.model is None:
+                        agent = agent_class()
+                    else:
+                        agent = agent_class(model=self.model)
                 else:
                     agent = agent_class_dict[row[agent_class_attr]](model=self.model)
 
@@ -172,13 +180,17 @@ class Creator:
 
         else:
             if n is not None:
-                agents = [agent_class(model=self.model) for _ in range(n)]
+                if self.model is None:
+                    agents = [agent_class() for _ in range(n)]
+                else:
+                    agents = [agent_class(model=self.model) for _ in range(n)]
+                
             else:
                 msg = "Either `df` or `n` must be not None."
                 raise Exception(msg)
-
-        agents = p2n.AgentList(model=self.model, objs=agents)
-
+        
+        # add agents to environment
+        self.env.add_agents(agents)
         return agents
 
     def _get_affiliated_agents(self, agents, dummy_location) -> list:
@@ -209,7 +221,7 @@ class Creator:
                     n_mother_locations_found += 1
 
             # Check if the number of mother locations is not 1
-            if n_mother_locations_found > 1 and self.model.enable_p2n_warnings:
+            if n_mother_locations_found > 1 and self.env.enable_p2n_warnings:
                 warnings.warn(
                     f"""For agent {agent},
                     {n_mother_locations_found} locations of class
@@ -331,7 +343,8 @@ class Creator:
                     #    # assign agents
                     for agent in sticky_agents:
                         group.append(agent)
-                        dummy_location.add_agent(agent)
+                        print(vars(dummy_location))
+                        #dummy_location.add_agent(agent)
 
                     assigned = True
                     break
@@ -343,7 +356,7 @@ class Creator:
                     # assign agents
[…]
