### 2025-05-01 79aafac
rm model from designer
