+++ src/pop2net/location.py
@@ -20,8 +20,12 @@ class Location:
         Args:
             model: Model this Location should be associated with.
         """
-        self.env = None
         self.label = self.__class__.__name__ if self.label is None else self.label
+        self.id_p2n = None
+        self.env_p2n = None
+    
+    def setup(self):
+        pass
 
     @property
     def agents(self) -> AgentList:
@@ -30,7 +34,7 @@ class Location:
         Returns:
             List of agents at this location.
         """
-        return self.env.agents_of_location(self)
+        return self.env_p2n.agents_of_location(self)
 
     def add_agent(self, agent: _agent.Agent, weight: float | None = None) -> None:
         """Assigns the given agent to this location.
@@ -41,7 +45,7 @@ class Location:
                 Defaults to None. If weight is set to None, the weight is generated by
                 location.weight(), which returns 1 by default.
         """
-        self.env.add_agent_to_location(self, agent=agent, weight=weight)
+        self.env_p2n.add_agent_to_location(self, agent=agent, weight=weight)
 
     def add_agents(self, agents: list, weight: float | None = None) -> None:
         """Add multiple agents at once.
@@ -62,7 +66,7 @@ class Location:
         Args:
             agent: Agent that is to be removed.
         """
-        self.env.remove_agent_from_location(
+        self.env_p2n.remove_agent_from_location(
             location=self,
             agent=agent,
         )
@@ -85,7 +89,7 @@ class Location:
         Returns:
             AgentList: A list of all agents at this location except the passed agent.
         """
-        agents = self.env.agents_of_location(self)
+        agents = self.env_p2n.agents_of_location(self)
         agents.remove(agent)
         return agents
 
@@ -98,7 +102,7 @@ class Location:
             agent (Agent): The agent.
             weight (float): The weight.
         """
-        self.env.set_weight(
+        self.env_p2n.set_weight(
             agent=agent,
             location=self,
             weight=weight,
@@ -113,7 +117,7 @@ class Location:
         Returns:
             Edge weight.
         """
-        return self.env.get_weight(agent=agent, location=self)
+        return self.env_p2n.get_weight(agent=agent, location=self)
 
     def weight(self, agent) -> float:  # noqa: ARG002
         """Generates the edge weight between a given agent and the location instance.
