### 2025-05-01 c0b5333
add attach_fresh_id
