### 2025-05-01 2017b70
rename model -> env and add framework-compatibility
