+++ src/pop2net/environment.py
@@ -6,7 +6,6 @@ import itertools
 import typing
 import warnings
 
-import agentpy as ap
 from agentpy import AgentList
 import networkx as nx
 
@@ -16,9 +15,11 @@ if typing.TYPE_CHECKING:
 
 from pop2net.sequences import LocationList
 
+from pop2net.creator import Creator
+from pop2net.inspector import NetworkInspector
 
-class Model:
-    """Class the encapsulates a full simluation.
+class Environment:
+    """Class that encapsulates a full simluation.
 
     This very closely follows the logic of the :class:`agentpy.Model` package. See
     :class:`agentpy.Model` for more information.
@@ -26,10 +27,9 @@ class Model:
 
     def __init__(
         self,
-        parameters=None,
-        _run_id=None,
+        model = None,
         enable_p2n_warnings=True,
-        **kwargs,
+        framework: str | None = None
     ):
         """Initiate a simulation.
 
@@ -40,14 +40,48 @@ class Model:
             Defaults to None.
             **kwargs: Optional parameters that are all passed to :class:`agentpy.Model`.
         """
-        self.g = nx.Graph()
+
+        # settings
+        self.framework: str | None = framework # TODO: if framework=="agentpy" return everything as AgentLists etc.
         self.enable_p2n_warnings = enable_p2n_warnings
+        
+        # import framework dependencies, if required
+        if self.framework is None:
+            pass
+        elif self.framework == "agentpy":
+            import agentpy
+            self._agentpy = agentpy
+        elif self.framework == "mesa":
+            import mesa
+            self._mesa = mesa
+        else:
+            raise ValueError(f"Invalid framework selected: {self.framework}")
+        
+        # connected objects
+        self.model = model
+        self.g = nx.Graph()
+        self.creator = Creator(env=self)
+        self.inspector = NetworkInspector(env=self)
+        
+        # dynamic attributes
         self._fresh_id = 0
-    
-    def attach_fresh_id(self, obj):
+
+    def _attach_fresh_id(self, obj):
         if obj.id_p2n is None:
             obj.id_p2n = self._fresh_id
             self._fresh_id += 1
+
+    def _to_framework(self, objects):
+        if self.framework is None:
+            return objects
+        elif self.framework == "agentpy":
+            return self._agentpy.AgentList(model=self.model, objs=objects)
+        elif self.framework == "mesa":
+            return self._mesa.agent.AgentSet(agents=objects, random=self.model.random)
+
+    def _get_objects(self, type_bipartite, nodes):
+        objects = [data["_obj"] for _, data in nodes if data["bipartite"] == type_bipartite]
+        return self._to_framework(objects=objects)
 
     @property
     def agents(self) -> list:
@@ -56,7 +90,7 @@ class Model:
         Returns:
             list: A non-mutable list of all agents in the environment.
         """
-        return [data["_obj"] for _, data in self.g.nodes(data=True) if data["bipartite"] == 0]
+        return self._get_objects(nodes=self.g.nodes(data=True), type_bipartite=0)
         
 
     @property
@@ -84,7 +118,8 @@ class Model:
         Returns:
             LocationList: a non-mutable LocationList of all locations in the environment.
         """
-        return [data["_obj"] for _, data in self.g.nodes(data=True) if data["bipartite"] == 1]
+        return self._get_objects(nodes=self.g.nodes(data=True), type_bipartite=1)
+        
 
     # TODO: def add_obj as a common parent method for add_agent & add_location
     def add_agent(self, agent: _agent.Agent) -> None:
@@ -97,11 +132,11 @@ class Model:
             agent: Agent to be added to the environment.
         """
         if agent.id_p2n is None:
-            self.attach_fresh_id(agent)
[…]
