+++ src/pop2net/location.py
@@ -6,7 +6,7 @@ from agentpy.objects import Object
 from agentpy.sequences import AgentList
 
 from . import agent as _agent
-from . import model as _model
+#from . import environment as _model
 
 
 class Location:
@@ -22,7 +22,7 @@ class Location:
         """
         self.label = self.__class__.__name__ if self.label is None else self.label
         self.id_p2n = None
-        self.env_p2n = None
+        self.env = None
     
     def setup(self):
         pass
@@ -34,7 +34,7 @@ class Location:
         Returns:
             List of agents at this location.
         """
-        return self.env_p2n.agents_of_location(self)
+        return self.env.agents_of_location(self)
 
     def add_agent(self, agent: _agent.Agent, weight: float | None = None) -> None:
         """Assigns the given agent to this location.
@@ -45,7 +45,7 @@ class Location:
                 Defaults to None. If weight is set to None, the weight is generated by
                 location.weight(), which returns 1 by default.
         """
-        self.env_p2n.add_agent_to_location(self, agent=agent, weight=weight)
+        self.env.add_agent_to_location(self, agent=agent, weight=weight)
 
     def add_agents(self, agents: list, weight: float | None = None) -> None:
         """Add multiple agents at once.
@@ -66,7 +66,7 @@ class Location:
         Args:
             agent: Agent that is to be removed.
         """
-        self.env_p2n.remove_agent_from_location(
+        self.env.remove_agent_from_location(
             location=self,
             agent=agent,
         )
@@ -89,7 +89,7 @@ class Location:
         Returns:
             AgentList: A list of all agents at this location except the passed agent.
         """
-        agents = self.env_p2n.agents_of_location(self)
+        agents = self.env.agents_of_location(self)
         agents.remove(agent)
         return agents
 
@@ -102,7 +102,7 @@ class Location:
             agent (Agent): The agent.
             weight (float): The weight.
         """
-        self.env_p2n.set_weight(
+        self.env.set_weight(
             agent=agent,
             location=self,
             weight=weight,
@@ -117,7 +117,7 @@ class Location:
         Returns:
             Edge weight.
         """
-        return self.env_p2n.get_weight(agent=agent, location=self)
+        return self.env.get_weight(agent=agent, location=self)
 
     def weight(self, agent) -> float:  # noqa: ARG002
         """Generates the edge weight between a given agent and the location instance.
