+++ pyproject.toml
@@ -32,7 +32,7 @@ license = "MIT"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4.0"
+python = ">=3.11,<4.0"
 numpy = [{version = ">=1.7", python = "<=3.12"}, {version = ">=1.26", python = ">=3.12"}]
 scipy = [{version = ">=1.6", python = "<=3.12"}, {version = ">=1.11.3", python = ">=3.12"}]
 pandas = ">=1.5.1"
@@ -41,6 +41,7 @@ tabulate = "^0.9.0"
 seaborn = "^0.12.1"
 faker = "^15.1.1"
 bokehgraph = "^1.1.0"
+mesa = "^3.1.5"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.2.0"
@@ -48,6 +49,7 @@ pre-commit = "^2.20.0"
 pytest-regressions = ">=2.4.1"
 ipython = "^8.7.0"
 ipykernel = "^6.17.1"
+mesa = "^3.0.0"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.1.0"
