### 2025-05-05 e6b15f0
remove  _get_objects() and return to old method to get agents etc.
