### 2025-05-05 2af226a
use agentpy only as a dev dependency
