+++ pyproject.toml
@@ -36,7 +36,6 @@ python = ">=3.11,<4.0"
 numpy = [{version = ">=1.7", python = "<=3.12"}, {version = ">=1.26", python = ">=3.12"}]
 scipy = [{version = ">=1.6", python = "<=3.12"}, {version = ">=1.11.3", python = ">=3.12"}]
 pandas = ">=1.5.1"
-agentpy = "^0.1.5"
 tabulate = "^0.9.0"
 seaborn = "^0.12.1"
 faker = "^15.1.1"
@@ -50,6 +49,7 @@ pytest-regressions = ">=2.4.1"
 ipython = "^8.7.0"
 ipykernel = "^6.17.1"
 mesa = "^3.0.0"
+agentpy = "^0.1.5"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.1.0"
