### 2025-05-05 abf333a
update a few tests to new api
