### 2025-05-06 2d57180
make location an agent, too
