### 2025-05-06 c4871d5
some more adjustments to make p2n compatible with both agentpy and mesa
