### 2025-05-06 c1cfc27
a first test for p2n with different frameworks (for mesa it is still failing)
