### 2025-05-07 #7897 @hugoledoux
I am back from a holidays. Thanks @jofmi and @EwoutH for the detailed comments 👍

One point of attention: we're really trying to keep the short papers <1000 words, because often they were getting too long. So while I agree with some of the points raised, JOSS favours a short paper with links to the docs.

So from the comments above: yes to explain what a bipartite network is (the paper should be understandable by non-experts), yes to putting the workflow, but perhaps no to putting code and several examples (one would be enough).

@mariuzka If you doubt about what to put and not, just ask me
