### 2025-05-07 83b8c62
add a test that tests p2n with agentpy and mesa
