+++ src/pop2net/inspector.py
@@ -29,9 +29,9 @@ class NetworkInspector:
 
     def plot_bipartite_network(
         self,
-        agent_attrs: list | None = None,
+        actor_attrs: list | None = None,
         location_attrs: list | None = None,
-        agent_color: str | None = None,
+        actor_color: str | None = None,
         location_color: str | None = None,
         edge_alpha: str = "weight",
         edge_color: str = "black",
@@ -39,15 +39,15 @@ class NetworkInspector:
         node_size: int = 10,
         node_alpha=0.5,
     ) -> None:
-        """Plots the bipartite network of agents and locations.
+        """Plots the bipartite network of actors and locations.
 
         Args:
-            agent_attrs (list | None, optional): A list of agent attributes that
+            actor_attrs (list | None, optional): A list of actor attributes that
                 should be shown as node attributes in the network graph. Defaults to None.
             location_attrs (list | None, optional): A list of location attributes that
                 should be shown as node attributes in the network graph. Defaults to None.
-            agent_color (str | None, optional): The agent attribute that determines the
-            color of the agent nodes. Defaults to None.
+            actor_color (str | None, optional): The actor attribute that determines the
+            color of the actor nodes. Defaults to None.
             location_color (str | None, optional): The location attribute that determines the
             color of the location nodes. Defaults to None.
             edge_alpha (str, optional): The edge attribute that determines the edges' transparency.
@@ -57,15 +57,15 @@ class NetworkInspector:
             node_size (int, optional): The size of the nodes. Defaults to 10.
             node_alpha (float, optional): The transparency of the nodes. Defaults to 0.5.
         """
-        #if agent_attrs is None:
-        #    agent_attrs = ["type"]
+        #if actor_attrs is None:
+        #    actor_attrs = ["type"]
         #else:
-        #    agent_attrs = list(agent_attrs)
-        #    if "type" not in agent_attrs:
-        #        agent_attrs.append("type")
+        #    actor_attrs = list(actor_attrs)
+        #    if "type" not in actor_attrs:
+        #        actor_attrs.append("type")
 
-        if agent_color is not None and agent_color not in agent_attrs:
-            agent_attrs.append(agent_color)
+        if actor_color is not None and actor_color not in actor_attrs:
+            actor_attrs.append(actor_color)
 
         #if location_attrs is None:
         #    location_attrs = ["type", "label"]
@@ -81,7 +81,7 @@ class NetworkInspector:
             location_attrs.append(location_color)
 
         graph = self.env.export_bipartite_network(
-            agent_attrs=agent_attrs,
+            actor_attrs=actor_attrs,
             location_attrs=location_attrs,
         )
 
@@ -97,7 +97,7 @@ class NetworkInspector:
 
         plot.draw(
             node_color=(
-                agent_color if agent_color is not None else "firebrick",
+                actor_color if actor_color is not None else "firebrick",
                 location_color if location_color is not None else "steelblue",
             ),
             edge_alpha=edge_alpha,
@@ -107,10 +107,10 @@ class NetworkInspector:
             node_alpha=node_alpha,
         )
 
-    def plot_agent_network(
+    def plot_actor_network(
         self,
-        agent_color: str | None = None,
-        agent_attrs: list | None = None,
+        actor_color: str | None = None,
+        actor_attrs: list | None = None,
         edge_alpha: str = "weight",
         edge_color: str = "black",
         include_0_weights: bool = True,
@@ -118,12 +118,12 @@ class NetworkInspector:
         node_size: int = 10,
         node_alpha=0.5,
     ) -> None:
-        """Plots the agent network.
+        """Plots the actor network.
 
         Args:
-            agent_color (str | None, optional): The agent attribute that determines the
-            color of the agent nodes. Defaults to None.
-            agent_attrs (list | None, optional): A list of agent attributes that
[…]
