+++ src/pop2net/creator.py
@@ -18,7 +18,7 @@ from .exceptions import Pop2netException
 
 
 class Creator:
-    """Creates and connects agents and locations."""
+    """Creates and connects actors and locations."""
 
     def __init__(
         self,
@@ -37,7 +37,7 @@ class Creator:
         self.rng = random.Random(seed)
         self.model = model
         self._dummy_model = type(self.model)()
-        self._temp_agent_attrs = ["_P2NTEMP_split_values", "_P2NTEMP_melt_location_weight"]
+        self._temp_actor_attrs = ["_P2NTEMP_split_values", "_P2NTEMP_melt_location_weight"]
 
     def _create_dummy_location(self, designer) -> p2n.Location:
         lc = designer.location_class
@@ -125,102 +125,102 @@ class Creator:
 
         return df_sample
 
-    def create_agents(
+    def create_actors(
         self,
-        agent_class=p2n.Agent,
-        agent_class_attr: None | str = None,
-        agent_class_dict: None | dict = None,
+        actor_class=p2n.Actor,
+        actor_class_attr: None | str = None,
+        actor_class_dict: None | dict = None,
         df: pd.DataFrame | None = None,
         n: int | None = None,
         clear: bool = False,
-    ) -> p2n.AgentList:
-        """Creates agents from a pandas DataFrame.
+    ):
+        """Creates actors from a pandas DataFrame.
 
-        Creates one agent-instance of the given agent-class for each row of the given df,
+        Creates one actor-instance of the given actor-class for each row of the given df,
         if df is not None.
         All columns of the df are added as instance attributes containing the row-specific values
         of the specific column.
-        If df is None and n is not None, n default agents without any additional attributes are
+        If df is None and n is not None, n default actors without any additional attributes are
         created.
 
         Args:
-            agent_class: A class to instantiate all agents with. Every column in the DataFrame will
-                result in an attribute of the agents.
-            df: The DataFrame from which the agents should be created from.
-            n: The number of agents that should be created. Defaults to None.
-            clear (bool): Should the agents already included in the model be removed?
+            actor_class: A class to instantiate all actors with. Every column in the DataFrame will
+                result in an attribute of the actors.
+            df: The DataFrame from which the actors should be created from.
+            n: The number of actors that should be created. Defaults to None.
+            clear (bool): Should the actors already included in the model be removed?
 
         Returns:
-            A list of agents.
+            A list of actors.
         """
         if clear:
-            self.env.remove_agents(self.env.agents)
+            self.env.remove_actors(self.env.actors)
 
         if df is not None:
             df = df.copy()
 
-            # create one agent for each row in df
-            agents = []
+            # create one actor for each row in df
+            actors = []
             for _, row in df.iterrows():
-                if agent_class_dict is None:
+                if actor_class_dict is None:
                     if self.model is None:
-                        agent = agent_class()
+                        actor = actor_class()
                     else:
-                        agent = agent_class(model=self.model)
+                        actor = actor_class(model=self.model)
                 else:
-                    agent = agent_class_dict[row[agent_class_attr]](model=self.model)
+                    actor = actor_class_dict[row[actor_class_attr]](model=self.model)
 
                 for col_name in df.columns:
                     if col_name == "id":
-                        msg = "You are not allowed to set an agent attribute called `id`."
+                        msg = "You are not allowed to set an actor attribute called `id`."
                         raise Exception(msg)
                     else:
-                        setattr(agent, col_name, row[col_name])
-                agents.append(agent)
+                        setattr(actor, col_name, row[col_name])
+                actors.append(actor)
 
         else:
             if n is not None:
[…]
