+++ src/pop2net/__init__.py
@@ -1,8 +1,6 @@
 """Pop2net. An extension package for location-based simulations based on AgentPy."""
 
-from agentpy import AgentList
-
-from .agent import Agent
+from .actor import Actor
 from .creator import Creator
 from .exceptions import Pop2netException
 from .inspector import NetworkInspector
@@ -10,11 +8,10 @@ from .location import Location
 from .location_designer import LocationDesigner
 from .location_designer import MeltLocationDesigner
 from .environment import Environment
-from .sequences import LocationList
+#from .sequences import LocationList
 
 __all__ = [
-    "AgentList",
-    "Agent",
+    "Actor",
     "Pop2netException",
     "Location",
     "LocationDesigner",

[…]
