+++ tests/test_agent/test_agent_connect_and_disconnect.py
@@ -1,175 +1,0 @@
-import pop2net as p2n
-
-
-def test_agent_connect_1():
-    env = p2n.Environment()
-    agent1 = p2n.Agent()
-    agent2 = p2n.Agent()
-    agent3 = p2n.Agent()
-    env.add_agents([agent1, agent2, agent3])
-
-    class Home(p2n.Location):
-        pass
-
-    class School(p2n.Location):
-        pass
-
-    agent1.connect(
-        agent=agent2,
-        location_cls=Home,
-    )
-
-    agent2.connect(
-        agent=agent3,
-        location_cls=School,
-    )
-
-    assert len(env.agents) == 3
-
-    assert isinstance(env.locations[0], Home)
-    assert isinstance(env.locations[1], School)
-
-    assert agent1 in env.locations[0].agents
-    assert agent1 not in env.locations[1].agents
-
-    assert agent2 in env.locations[0].agents
-    assert agent2 in env.locations[1].agents
-
-    assert agent3 not in env.locations[0].agents
-    assert agent3 in env.locations[1].agents
-
-    agent1.disconnect(
-        agent2,
-        location_labels=None,
-        remove_locations=False,
-        remove_neighbor=True,
-        remove_self=True,
-    )
-
-
-def test_agent_disconnect_1():
-    env = p2n.Environment()
-    agent1 = p2n.Agent()
-    agent2 = p2n.Agent()
-    agent3 = p2n.Agent()
-    env.add_agents([agent1, agent2, agent3])
-
-    class Home(p2n.Location):
-        pass
-
-    class School(p2n.Location):
-        pass
-
-    agent1.connect(
-        agent=agent2,
-        location_cls=Home,
-    )
-
-    agent2.connect(
-        agent=agent3,
-        location_cls=School,
-    )
-
-    agent1.disconnect(
-        agent2,
-        location_labels=None,
-        remove_locations=False,
-        remove_neighbor=True,
-        remove_self=True,
-    )
-
-    assert len(env.agents) == 3
-
-    assert isinstance(env.locations[0], Home)
-    assert isinstance(env.locations[1], School)
-
-    assert agent1 not in env.locations[0].agents
-    assert agent1 not in env.locations[1].agents
-
-    assert agent2 not in env.locations[0].agents
-    assert agent2 in env.locations[1].agents
-
-    assert agent3 not in env.locations[0].agents
-    assert agent3 in env.locations[1].agents
-
-    agent2.disconnect(
-        agent3,
-        location_labels=None,
-        remove_locations=False,
-        remove_neighbor=True,
-        remove_self=False,
-    )
-
-    assert len(env.agents) == 3
-
-    assert isinstance(env.locations[0], Home)
-    assert isinstance(env.locations[1], School)
-
-    assert agent1 not in env.locations[0].agents
-    assert agent1 not in env.locations[1].agents
-
-    assert agent2 not in env.locations[0].agents
-    assert agent2 in env.locations[1].agents
-
-    assert agent3 not in env.locations[0].agents
-    assert agent3 not in env.locations[1].agents
-
-
-def test_agent_disconnect_2():
-    env = p2n.Environment()
-    agent1 = p2n.Agent()
-    agent2 = p2n.Agent()
-    agent3 = p2n.Agent()
-    env.add_agents([agent1, agent2, agent3])
-
-    class Home(p2n.Location):
-        pass
-
-    class School(p2n.Location):
-        pass
-
-    agent1.connect(
-        agent=agent2,
-        location_cls=Home,
-    )
-
-    agent1.connect(
-        agent=agent2,
-        location_cls=School,
-    )
-
-    agent1.disconnect(
-        agent2,
-        location_labels=["Home"],
[…]
