### 2025-05-07 f47cb0d
update docstrings and type hints
