+++ src/pop2net/actor.py
@@ -20,7 +20,7 @@ class Actor:
         self.env = None
         self.id_p2n = None
         self.model = None
-        self.type = "Actor"
+        self.label = self.__class__.__name__ if self.label is None else self.label
         super().__init__(*args, **kwargs)
 
     def neighbors(self, location_labels: list[str] | None = None) -> list:
