+++ src/pop2net/environment.py
@@ -12,8 +12,6 @@ if typing.TYPE_CHECKING:
     from . import actor as _actor
     from . import location as _location
 
-from pop2net.creator import Creator
-from pop2net.inspector import NetworkInspector
 import pop2net as p2n
 
 
@@ -55,8 +53,8 @@ class Environment:
         self.g = nx.Graph()
 
         # TODO: should we add creator and inspector as attributes?
-        #self.creator = Creator(env=self)
-        #self.inspector = NetworkInspector(env=self)
+        # self.creator = Creator(env=self)
+        # self.inspector = NetworkInspector(env=self)
 
         # a unique id that is added to the objects of this environment
         self._fresh_id = 0
