### 2025-05-08 3ca9620
make location_designer compatible
