### 2025-05-08 97a9815
shared location test rewritten
TODO: Version with location_designer?
