### 2025-05-08 6a7240c
auto-commit: fixes by ruff
