+++ tests/test_agent/test_agent_shared_locations.py
@@ -1,5 +1,6 @@
 import pop2net as p2n
 
+
 def test_1():
     env = p2n.Environment()
     actor1 = p2n.Actor()
@@ -13,41 +14,17 @@ def test_1():
     class School(p2n.Location):
         pass
 
-    actor1.connect(
-        actor=actor2,
-        location_cls=Home
-    )
+    actor1.connect(actor=actor2, location_cls=Home)
 
-    actor2.connect(
-        actor=actor3,
-        location_cls=School
-    )
+    actor2.connect(actor=actor3, location_cls=School)
 
-    actor1_2_shared = actor1.shared_locations(
-        actor=actor2,
-        location_labels=None
-    )
-    actor2_1_shared = actor2.shared_locations(
-        actor=actor1,
-        location_labels=None
-    )
-    actor2_3_shared = actor2.shared_locations(
-        actor=actor3,
-        location_labels=None
-    )
-    actor3_2_shared = actor3.shared_locations(
-        actor=actor2,
-        location_labels=None
-    )
-    actor1_3_shared = actor1.shared_locations(
-        actor=actor3,
-        location_labels=None
-    )
-    actor3_1_shared = actor3.shared_locations(
-        actor=actor1,
-        location_labels=None
-    )
-    
+    actor1_2_shared = actor1.shared_locations(actor=actor2, location_labels=None)
+    actor2_1_shared = actor2.shared_locations(actor=actor1, location_labels=None)
+    actor2_3_shared = actor2.shared_locations(actor=actor3, location_labels=None)
+    actor3_2_shared = actor3.shared_locations(actor=actor2, location_labels=None)
+    actor1_3_shared = actor1.shared_locations(actor=actor3, location_labels=None)
+    actor3_1_shared = actor3.shared_locations(actor=actor1, location_labels=None)
+
     assert len(env.actors) == 3
 
     assert isinstance(env.locations[0], Home)
@@ -70,10 +47,10 @@ def test_1():
     assert len(actor1_3_shared) == 0
     assert len(actor3_1_shared) == 0
 
-    assert (isinstance(actor1_2_shared[0], Home))
-    assert (isinstance(actor2_1_shared[0], Home))
-    assert (isinstance(actor2_3_shared[0], School))
-    assert (isinstance(actor3_2_shared[0], School))
+    assert isinstance(actor1_2_shared[0], Home)
+    assert isinstance(actor2_1_shared[0], Home)
+    assert isinstance(actor2_3_shared[0], School)
+    assert isinstance(actor3_2_shared[0], School)
 
 
 # TODO Location_designer version when its ready
