+++ src/pop2net/environment.py
@@ -420,6 +420,7 @@ class Environment:
             if self.framework is None:
                 location_cls = p2n.Location
             else:
+
                 class Location(p2n.Location, self._framework.Agent):
                     pass
 
