### 2025-05-08 c666713
test_1b updated with new syntax
