### 2025-05-11 e45df57
location_label and actor_neighbors tests updated for new syntax
