+++ tests/test_agent/test_agent_neighbors.py
@@ -1,84 +1,87 @@
+#%%
 import pop2net as p2n
 
 
 def test_1():
     env = p2n.Environment()
-    agent1 = p2n.Agent()
-    agent2 = p2n.Agent()
+    actor1 = p2n.Actor()
+    actor2 = p2n.Actor()
     location1 = p2n.Location()
 
-    env.add_agents([agent1, agent2])
+    env.add_actors([actor1, actor2])
     env.add_location(location1)
 
-    location1.add_agents([agent1, agent2])
+    location1.add_actors([actor1, actor2])
 
-    assert len(env.agents) == 2
+    assert len(env.actors) == 2
     assert len(env.locations) == 1
 
-    assert len(agent1.neighbors()) == 1
-    assert agent1.neighbors()[0] is agent2
+    assert len(actor1.neighbors()) == 1
+    assert actor1.neighbors()[0] is actor2
 
-    assert len(agent2.neighbors()) == 1
-    assert agent2.neighbors()[0] is agent1
+    assert len(actor2.neighbors()) == 1
+    assert actor2.neighbors()[0] is actor1
 
 
 def test_2a():
-    class Max(p2n.Agent):
+    class Max(p2n.Actor):
         pass
 
-    class Marius(p2n.Agent):
+    class Marius(p2n.Actor):
         pass
 
-    class Lukas(p2n.Agent):
+    class Lukas(p2n.Actor):
         pass
 
     class WebexMeeting(p2n.LocationDesigner):
         pass
 
-    model = p2n.Model()
-    agent_max = Max(model=model)
-    agent_marius = Marius(model=model)
-    agent_lukas = Lukas(model=model)
-    meeting = WebexMeeting(model=model)
-    meeting.add_agents(agents=[agent_max, agent_marius, agent_lukas])
+    env = p2n.Environment()
+    actor_max = Max()
+    actor_marius = Marius()
+    actor_lukas = Lukas()
+    meeting = WebexMeeting()
+    env.add_actors(actors=[actor_max, actor_marius, actor_lukas])
+    env.add_location(meeting)
+    meeting.add_actors(actors=[actor_max, actor_marius, actor_lukas])
 
-    assert len(model.locations) == 1
-    assert len(model.agents) == 3
+    assert len(env.locations) == 1
+    assert len(env.actors) == 3
 
-    assert agent_max.neighbors()[0].type == "Marius"
-    assert agent_max.neighbors()[1].type == "Lukas"
+    assert actor_max.neighbors()[0].type == "Marius"
+    assert actor_max.neighbors()[1].type == "Lukas"
 
-    assert agent_marius.neighbors()[0].type == "Max"
-    assert agent_marius.neighbors()[1].type == "Lukas"
+    assert actor_marius.neighbors()[0].type == "Max"
+    assert actor_marius.neighbors()[1].type == "Lukas"
 
-    assert agent_lukas.neighbors()[0].type == "Max"
-    assert agent_lukas.neighbors()[1].type == "Marius"
+    assert actor_lukas.neighbors()[0].type == "Max"
+    assert actor_lukas.neighbors()[1].type == "Marius"
 
 
 # all in one location
 def test_2b():
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
-    class Max(p2n.Agent):
+    class Max(p2n.Actor):
         pass
 
-    class Marius(p2n.Agent):
+    class Marius(p2n.Actor):
         pass
 
-    class Lukas(p2n.Agent):
+    class Lukas(p2n.Actor):
         pass
 
     class WebexMeeting(p2n.LocationDesigner):
         pass
 
[…]
