### 2025-05-12 72e4285
show label/type in network plot again
