+++ tests/test_agent/test_agent_neighbors.py
@@ -1,4 +1,4 @@
-#%%
+# %%
 import pop2net as p2n
 
 
