### 2025-05-13 c505bac
partially upgraded with new syntax, but still not working with soep
