### 2025-05-15 ca376c9
all test_creator updated
Problem with test_creator_delete_magic_attrs.py. Could not make it to run through, but some different error than assert
