+++ tests/test_creator/test_creator_delete_magic_attrs.py
@@ -5,17 +5,17 @@ import pop2net as p2n
 
 
 @pytest.fixture(scope="session")
-def magic_agent_attributes():
-    magic_agent_attributes = ["", "_assigned", "_id", "_position", "_head", "_tail"]
-    magic_agent_attributes = ["Location1" + attr for attr in magic_agent_attributes] + [
-        "Location2" + attr for attr in magic_agent_attributes
+def magic_actor_attributes():
+    magic_actor_attributes = ["", "_assigned", "_id", "_position", "_head", "_tail"]
+    magic_actor_attributes = ["Location1" + attr for attr in magic_actor_attributes] + [
+        "Location2" + attr for attr in magic_actor_attributes
     ]
-    return magic_agent_attributes
+    return magic_actor_attributes
 
 
-def test_del_magic_agent_attrs1(magic_agent_attributes):
-    model = p2n.Model()
-    creator = p2n.Creator(model=model)
+def test_del_magic_agent_attrs1(magic_actor_attributes):
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class Location1(p2n.LocationDesigner):
         pass
@@ -24,34 +24,34 @@ def test_del_magic_agent_attrs1(magic_agent_attributes):
         pass
 
     # First test: Do NOT delete magic agent attributes
-    creator.create_agents(n=10)
+    creator.create_actors(n=10)
     creator.create_locations(
         location_designers=[Location1, Location2],
-        delete_magic_agent_attributes=False,
+        delete_magic_actor_attributes=False,
     )
 
-    for agent in model.agents:
-        for attr in magic_agent_attributes:
-            assert hasattr(agent, attr)
+    for actor in env.actors:
+        for attr in magic_actor_attributes:
+            assert hasattr(actor, attr)
 
     # Second test: DELETE magic agent attributes
-    creator.create_agents(n=10, clear=True)
+    creator.create_actors(n=10, clear=True)
     creator.create_locations(
         location_designers=[Location1, Location2],
-        delete_magic_agent_attributes=True,
+        delete_magic_actor_attributes=True,
         clear=True,
     )
 
-    for agent in model.agents:
-        for attr in magic_agent_attributes:
-            assert not hasattr(agent, attr)
+    for agent in env.actors:
+        for attr in magic_actor_attributes:
+            assert not hasattr(actor, attr)
 
 
-def test_del_magic_agent_attrs2(magic_agent_attributes):
+def test_del_magic_agent_attrs2(magic_actor_attributes):
     df = pd.DataFrame({"testattr": [10, 11, 12, 13]})
 
-    model = p2n.Model()
-    creator = p2n.Creator(model=model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class Location1(p2n.LocationDesigner):
         pass
@@ -62,24 +62,24 @@ def test_del_magic_agent_attrs2(magic_agent_attributes):
     # First test: Do NOT delete magic agent attributes
     creator.create(
         df=df,
-        n_agents=10,
+        n_actors=10,
         location_designers=[Location1, Location2],
-        delete_magic_agent_attributes=False,
+        delete_magic_actor_attributes=False,
     )
 
-    for agent in model.agents:
-        for attr in magic_agent_attributes:
-            assert hasattr(agent, attr)
+    for actor in env.actors:
+        for attr in magic_actor_attributes:
+            assert hasattr(actor, attr)
 
     # Second test: DELETE magic agent attributes
     creator.create(
         df=df,
         n_agents=10,
         location_designers=[Location1, Location2],
-        delete_magic_agent_attributes=True,
+        delete_magic_actor_attributes=True,
         clear=True,
     )
 
[…]
