### 2025-05-15 04345c3
updated syntax for all tests in test_combinations
