+++ tests/test_location_designer/test_combinations/test_n_agents_split.py
@@ -10,27 +10,27 @@ def test_1():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        n_agents = 2
+        n_actors = 2
 
-        def split(self, agent):
-            return agent.status
+        def split(self, actor):
+            return actor.status
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 5
-    assert len(model.agents) == 10
-    for location in model.locations:
-        if location.agents[0].status == "A":
-            assert len(location.agents) == 2
-            assert all(agent.status == "A" for agent in location.agents)
-        if location.agents[0].status == "B":
-            assert len(location.agents) == 2
-            assert all(agent.status == "B" for agent in location.agents)
+    assert len(env.locations) == 5
+    assert len(env.actors) == 10
+    for location in env.locations:
+        if location.actors[0].status == "A":
+            assert len(location.actors) == 2
+            assert all(actor.status == "A" for actor in location.actors)
+        if location.actors[0].status == "B":
+            assert len(location.actors) == 2
+            assert all(actor.status == "B" for actor in location.actors)
 
 
 def test_2():
@@ -41,28 +41,28 @@ def test_2():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        n_agents = 1
+        n_actors = 1
 
-        def split(self, agent):
-            return [agent.status, agent.sex]
+        def split(self, actor):
+            return [actor.status, actor.sex]
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 8
-    assert len(model.agents) == 4
-    assert len(model.locations[0].agents) == 1
-    assert len(model.locations[1].agents) == 1
-    assert len(model.locations[2].agents) == 1
-    assert len(model.locations[3].agents) == 1
-    assert len(model.locations[4].agents) == 1
-    assert len(model.locations[5].agents) == 1
-    assert len(model.locations[6].agents) == 1
-    assert len(model.locations[7].agents) == 1
+    assert len(env.locations) == 8
+    assert len(env.actors) == 4
+    assert len(env.locations[0].actors) == 1
+    assert len(env.locations[1].actors) == 1
+    assert len(env.locations[2].actors) == 1
+    assert len(env.locations[3].actors) == 1
+    assert len(env.locations[4].actors) == 1
+    assert len(env.locations[5].actors) == 1
+    assert len(env.locations[6].actors) == 1
+    assert len(env.locations[7].actors) == 1
 
 
 def test_3():
@@ -73,21 +73,21 @@ def test_3():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        n_agents = 1
+        n_actors = 1
 
-        def split(self, agent):
-            return str(agent.status) + str(agent.sex)
+        def split(self, actor):
+            return str(actor.status) + str(actor.sex)
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 4
[…]
