### 2025-05-15 068a473
updated syntax for test_location tests (1 test)
