+++ tests/test_location/test_location.py
@@ -1,4 +1,4 @@
-from pop2net.agent import Agent
+from pop2net.actor import Actor
 from pop2net.environment import Environment
 from pop2net.location import Location
 
@@ -8,26 +8,26 @@ def test_location():
 
     loc = Location()
 
-    agent1 = Agent()
-    agent2 = Agent()
+    actor1 = Actor()
+    actor2 = Actor()
 
     env.add_location(loc)
-    env.add_agents([agent1, agent2])
+    env.add_actors([actor1, actor2])
 
-    assert loc.agents == []
+    assert loc.actors == []
 
-    loc.add_agent(agent1, weight=1)
+    loc.add_actor(actor1, weight=1)
 
-    assert loc.agents == [agent1]
+    assert loc.actors == [actor1]
 
-    loc.add_agent(agent2, weight=1)
+    loc.add_actor(actor2, weight=1)
 
-    assert loc.agents == [agent1, agent2]
+    assert loc.actors == [actor1, actor2]
 
-    loc.remove_agent(agent1)
+    loc.remove_actor(actor1)
 
-    assert loc.agents == [agent2]
+    assert loc.actors == [actor2]
 
-    loc.remove_agent(agent2)
+    loc.remove_actor(actor2)
 
-    assert loc.agents == []
+    assert loc.actors == []
