### 2025-05-15 37b0497
upgraded syntax for tests in test_location_designer
