+++ tests/test_location_designer/test_mutate_and_multi_membership.py
@@ -2,38 +2,38 @@ import pop2net as p2n
 
 
 def test_1():
-    """Tests if mutate() can be used to create locations with varying number of agents while
-    ensuring that agents are only assigned to one mutation of the location designer."""
+    """Tests if mutate() can be used to create locations with varying number of actors while
+    ensuring that actors are only assigned to one mutation of the location designer."""
 
-    class TestAgent(p2n.Agent):
-        def __init__(self, model):
-            super().__init__(model)
+    class TestActor(p2n.Actor):
+        def __init__(self, env):
+            super().__init__(env)
 
     class TestLocation(p2n.LocationDesigner):
         n_locations = 1
 
-        def filter(self, agent):
-            """Prevent agents from being assigned to multiple locations of this type."""
-            if not any(label.startswith("TestLocation") for label in agent.location_labels):
+        def filter(self, actor):
+            """Prevent actors from being assigned to multiple locations of this type."""
+            if not any(label.startswith("TestLocation") for label in actor.location_labels):
                 return True
             else:
                 return False
 
         def mutate(self):
-            """Create multiple locations with varying number of agents."""
-            return {"n_agents": [2, 2, 2, 4]}
+            """Create multiple locations with varying number of actors."""
+            return {"n_actors": [2, 2, 2, 4]}
 
-    model = p2n.Model()
-    creator = p2n.Creator(model=model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     for _ in range(10):
-        TestAgent(model=model)
+        TestActor(env=env)
 
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.agents) == 10
-    assert len(model.locations) == 4
-    assert len(model.locations[0].agents) == 2
-    assert len(model.locations[1].agents) == 2
-    assert len(model.locations[2].agents) == 2
-    assert len(model.locations[3].agents) == 4
+    assert len(env.actors) == 10
+    assert len(env.locations) == 4
+    assert len(env.locations[0].actors) == 2
+    assert len(env.locations[1].actors) == 2
+    assert len(env.locations[2].actors) == 2
+    assert len(env.locations[3].actors) == 4
