+++ tests/test_location_designer/test_location_stick_together.py
@@ -7,58 +7,58 @@ def test_1():
     df = pd.DataFrame({"status": ["pupil", "pupil", "pupil", "pupil"], "class_id": [1, 2, 1, 2]})
 
     class Classroom(p2n.LocationDesigner):
-        n_agents = 2
+        n_actors = 2
 
-        def stick_together(self, agent):
-            return agent.class_id
+        def stick_together(self, actor):
+            return actor.class_id
 
-    model = p2n.Model()
-    creator = p2n.Creator(model=model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
     creator.create(df=df, location_designers=[Classroom])
 
-    assert len(model.agents) == 4
-    assert len(model.locations) == 2
+    assert len(env.actors) == 4
+    assert len(env.locations) == 2
 
-    for location in model.locations:
-        assert len(location.agents) == 2
+    for location in env.locations:
+        assert len(location.actors) == 2
 
-    for agent in model.agents:
-        assert agent.neighbors(location_labels=["Classroom"])[0].class_id == agent.class_id
+    for actor in env.actors:
+        assert actor.neighbors(location_labels=["Classroom"])[0].class_id == actor.class_id
 
-    inspector = p2n.NetworkInspector(model)
+    inspector = p2n.NetworkInspector(env)
     inspector.plot_bipartite_network()
-    inspector.plot_agent_network(agent_attrs=df.columns, agent_color="class_id")
+    inspector.plot_actor_network(actor_attrs=df.columns, actor_color="class_id")
 
 
-# stick together with uneven agent-location "seats"
+# stick together with uneven actor-location "seats"
 def test_2():
     df = pd.DataFrame(
         {"status": ["pupil", "pupil", "pupil", "pupil", "pupil"], "class_id": [1, 2, 1, 2, 1]}
     )
 
     class Classroom(p2n.LocationDesigner):
-        n_agents = 2
+        n_actors = 2
 
-        def stick_together(self, agent):
-            return agent.class_id
+        def stick_together(self, actor):
+            return actor.class_id
 
-    model = p2n.Model()
-    creator = p2n.Creator(model=model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
     creator.create(df=df, location_designers=[Classroom])
-    inspector = p2n.NetworkInspector(model)
+    inspector = p2n.NetworkInspector(env)
     inspector.plot_bipartite_network()
-    inspector.plot_agent_network(agent_attrs=df.columns, agent_color="class_id")
+    inspector.plot_actor_network(actor_attrs=df.columns, actor_color="class_id")
 
-    assert len(model.agents) == 5
-    assert len(model.locations) == 2
+    assert len(env.actors) == 5
+    assert len(env.locations) == 2
 
     expected_loc_lens = [2, 3]
-    for location in model.locations:
-        assert len(location.agents) in expected_loc_lens
-        del expected_loc_lens[expected_loc_lens.index(len(location.agents))]
+    for location in env.locations:
+        assert len(location.actors) in expected_loc_lens
+        del expected_loc_lens[expected_loc_lens.index(len(location.actors))]
 
-    for agent in model.agents:
-        assert all(nghbr.class_id == agent.class_id for nghbr in agent.neighbors())
+    for actor in env.actors:
+        assert all(nghbr.class_id == actor.class_id for nghbr in actor.neighbors())
 
 
 # stick_together with split
@@ -72,38 +72,38 @@ def test_3():
     )
 
     class Classroom(p2n.LocationDesigner):
-        n_agents = 2
+        n_actors = 2
 
-        def split(self, agent):
-            return agent.class_id
+        def split(self, actor):
+            return actor.class_id
 
[…]
