+++ tests/test_location_designer/test_location_weight.py
@@ -10,27 +10,27 @@ def test_1():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        def weight(self, agent):
+        def weight(self, actor):
             return 1
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 1
-    assert len(model.agents) == 5
-    assert len(model.locations[0].agents) == 5
+    assert len(env.locations) == 1
+    assert len(env.actors) == 5
+    assert len(env.locations[0].actors) == 5
     # sum of weights correct?
     assert (
-        sum([model.get_weight(agent, model.locations[0]) for agent in model.locations[0].agents])
+        sum([env.get_weight(actor, env.locations[0]) for actor in env.locations[0].actors])
         == 5
     )
-    # inidividual weights correct?
+    # individual weights correct?
     assert all(
-        model.get_weight(agent, model.locations[0]) == 1 for agent in model.locations[0].agents
+        env.get_weight(actor, env.locations[0]) == 1 for actor in env.locations[0].actors
     )
 
 
@@ -41,33 +41,33 @@ def test_2():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        def weight(self, agent):
-            if agent.status == "A":
+        def weight(self, actor):
+            if actor.status == "A":
                 return 1
-            if agent.status == "B":
+            if actor.status == "B":
                 return 5
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 1
-    assert len(model.agents) == 5
-    assert len(model.locations[0].agents) == 5
+    assert len(env.locations) == 1
+    assert len(env.actors) == 5
+    assert len(env.locations[0].actors) == 5
     # sum of weights correct?
     assert (
-        sum([model.get_weight(agent, model.locations[0]) for agent in model.locations[0].agents])
+        sum([env.get_weight(actor, env.locations[0]) for actor in env.locations[0].actors])
         == 17
     )
-    # inidividual weights correct?
+    # individual weights correct?
     assert all(
-        model.get_weight(agent, model.locations[0]) == 1
-        if agent.status == "A"
-        else model.get_weight(agent, model.locations[0]) == 5
-        for agent in model.locations[0].agents
+        env.get_weight(actor, env.locations[0]) == 1
+        if actor.status == "A"
+        else env.get_weight(actor, env.locations[0]) == 5
+        for actor in env.locations[0].actors
     )
 
 
@@ -78,28 +78,28 @@ def test_3():
         },
     )
 
-    model = p2n.Model()
-    creator = p2n.Creator(model)
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
 
     class TestLocation(p2n.LocationDesigner):
-        def weight(self, agent):
-            return agent.status
+        def weight(self, actor):
+            return actor.status
 
-    creator.create_agents(df=df)
+    creator.create_actors(df=df)
     creator.create_locations(location_designers=[TestLocation])
 
-    assert len(model.locations) == 1
-    assert len(model.agents) == 5
-    assert len(model.locations[0].agents) == 5
+    assert len(env.locations) == 1
+    assert len(env.actors) == 5
+    assert len(env.locations[0].actors) == 5
[…]

[…]
