### 2025-05-15 a6d0a5e
upgraded syntax for test_model tests
