+++ tests/test_model/test_model_connect_and_disconnect.py
@@ -1,67 +1,14 @@
 import pop2net as p2n
 
 
-def test_model_connect_agents_and_disconnect_agents_1():
-    """A test without specifing locationt types and without removing locations from model."""
+def test_model_connect_actors_and_disconnect_actors_1():
+    """A test without specifying location types and without removing locations from env."""
 
-    model = p2n.Model()
-    agent1 = p2n.Agent(model)
-    agent2 = p2n.Agent(model)
-    agent3 = p2n.Agent(model)
-    agents = [agent1, agent2, agent3]
-
-    class Home(p2n.Location):
-        pass
-
-    class School(p2n.Location):
-        pass
-
-    model.connect_agents(
-        agents=agents,
-        location_cls=Home,
-    )
-
-    model.connect_agents(
-        agents=[agent1, agent2],
-        location_cls=School,
-    )
-
-    assert len(model.agents) == 3
-
-    assert isinstance(model.locations[0], Home)
-    assert isinstance(model.locations[1], School)
-
-    assert agent1 in model.locations[0].agents
-    assert agent1 in model.locations[1].agents
-    assert agent2 in model.locations[0].agents
-    assert agent2 in model.locations[1].agents
-    assert agent3 in model.locations[0].agents
-    assert agent3 not in model.locations[1].agents
-
-    model.disconnect_agents(agents, remove_locations=False)
-
-    assert len(model.agents) == 3
-
-    assert len(model.locations) == 2
-    assert isinstance(model.locations[0], Home)
-    assert isinstance(model.locations[1], School)
-
-    assert len(model.locations[0].agents) == 0
-    assert len(model.locations[1].agents) == 0
-
-    assert len(agent1.locations) == 0
-    assert len(agent2.locations) == 0
-    assert len(agent3.locations) == 0
-
-
-def test_model_connect_agents_and_disconnect_agents_2():
-    """A test without removing locations from model."""
-
-    model = p2n.Model()
-    agent1 = p2n.Agent(model)
-    agent2 = p2n.Agent(model)
-    agent3 = p2n.Agent(model)
-    agents = [agent1, agent2, agent3]
+    env = p2n.Environment()
+    actor1 = p2n.Actor(env)
+    actor2 = p2n.Actor(env)
+    actor3 = p2n.Actor(env)
+    actors = [actor1, actor2, actor3]
 
     class Home(p2n.Location):
         pass
@@ -69,71 +16,52 @@ def test_model_connect_agents_and_disconnect_agents_2():
     class School(p2n.Location):
         pass
 
-    model.connect_agents(
-        agents=agents,
+    env.connect_actors(
+        actors=actors,
         location_cls=Home,
     )
 
-    model.connect_agents(
-        agents=[agent1, agent2],
+    env.connect_actors(
+        actors=[actor1, actor2],
         location_cls=School,
     )
 
-    model.disconnect_agents(agents, location_labels=["Home"], remove_locations=False)
+    assert len(env.actors) == 3
 
-    assert len(model.agents) == 3
+    assert isinstance(env.locations[0], Home)
+    assert isinstance(env.locations[1], School)
 
-    assert len(model.locations) == 2
-    assert isinstance(model.locations[0], Home)
-    assert isinstance(model.locations[1], School)
+    assert actor1 in env.locations[0].actors
+    assert actor1 in env.locations[1].actors
+    assert actor2 in env.locations[0].actors
+    assert actor2 in env.locations[1].actors
+    assert actor3 in env.locations[0].actors
+    assert actor3 not in env.locations[1].actors
 
-    assert len(model.locations[0].agents) == 0
[…]
