+++ tests/test_location_designer/test_combinations/test_n_agents_nest.py
@@ -171,9 +171,7 @@ def test_3():
         assert location.actors[0].group == location.actors[1].group
 
     # not all members of the same group are also in the same city (which is not desired)
-    assert not all(
-        location.actors[0].City == location.actors[1].City for location in env.locations
-    )
+    assert not all(location.actors[0].City == location.actors[1].City for location in env.locations)
 
     class GroupNestedInCity(Group):
         def nest(self):
