+++ tests/test_location_designer/test_location_nest_marius.py
@@ -38,9 +38,7 @@ def test_1():
         assert location.actors[0].group == location.actors[1].group
 
     # not all members of the same group are also in the same city (which is not desired)
-    assert not all(
-        location.actors[0].City == location.actors[1].City for location in env.locations
-    )
+    assert not all(location.actors[0].City == location.actors[1].City for location in env.locations)
 
     class GroupNestedInCityDesigner(GroupDesigner):
         def nest(self):
