### 2025-05-19 5ceafe8
updated additional syntax changes
- removed params of instantiations (e.g. Actor(env))
- added env.add_actors/locations
- some other fixes
