+++ tests/test_location_designer/test_combinations/test_split_stick_together.py
@@ -11,7 +11,7 @@ def test_1():
     df = pd.DataFrame(
         {
             "friend_group": [1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4],
-            "split_group": [1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2],
+            "split_group": [1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2, 1, 2],
         }
     )
 
