### 2025-05-19 930e893
auto-commit: fixes by ruff
