+++ tests/test_model/test_model_agents_and_locations_by_id.py
@@ -1,24 +1,0 @@
-import pop2net as p2n
-
-
-def test_1():
-    env = p2n.Environment()
-
-    env.add_actors([p2n.Actor() for _ in range(10)])
-    env.add_actors([p2n.Location() for _ in range(10)])
-    env.add_actors([p2n.Actor() for _ in range(10)])
-    env.add_actors([p2n.Location() for _ in range(10)])
-
-    # check if the actors can be found by id actors_dict
-    for actor in env.actors:
-        assert actor is env.actors_by_id[actor.id]
-
-    # check if the locations can be found by id locations_dict
-    for location in env.locations:
-        assert location is env.locations_by_id[location.id]
-
-    # assert that the actors cannot be found by id in the normal actor_list
-    assert not all(actor is env.actors[actor.id] for actor in env.actors)
-
-    # assert that the locations cannot be found by id in the normal locations_list
-    assert not all(actor is env.actors[actor.id] for actor in env.actors)
