### 2025-05-20 f96c257
fix test_model
