### 2025-05-20 c57679c
rm inspectors from tests
