+++ tests/test_location_designer/test_location_nest_marius.py
@@ -8,7 +8,7 @@ def test_1():
 
     class GroupDesigner(p2n.LocationDesigner):
         label = "Group"
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -39,7 +39,7 @@ def test_1():
         assert location.actors[0].group == location.actors[1].group
 
     # not all members of the same group are also in the same city (which is not desired)
-    assert not all(location.actors[0].City == location.actors[1].City for location in env.locations)
+    assert not all(location.actors[0].City == location.actors[-1].City for location in env.locations)
 
     class GroupNestedInCityDesigner(GroupDesigner):
         def nest(self):
