+++ tests/test_location_designer/test_location_nest.py
@@ -58,7 +58,7 @@ def test_2():
 
     class ClassroomDesigner(p2n.LocationDesigner):
         label = "Classroom"
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -72,7 +72,7 @@ def test_2():
     )
 
     assert len(env.actors) == 8
-    assert len(env.locations) == 6
+    assert len(env.locations) == 4
 
     for location in env.locations:
         if location.label == "School":
@@ -82,7 +82,7 @@ def test_2():
             assert counter[2] == 2
 
     assert not all(
-        location.actors[0].School == location.actors[1].School
+        location.actors[0].School == location.actors[-1].School
         for location in env.locations
         if location.label == "Classroom"
     )
@@ -93,7 +93,7 @@ def test_2():
 
     class ClassroomDesigner(p2n.LocationDesigner):
         label = "Classroom"
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -112,7 +112,7 @@ def test_2():
     assert len(env.actors) == 8
     assert len(env.locations) == 6
     assert all(
-        location.actors[0].School == location.actors[1].School
+        location.actors[0].School == location.actors[-1].School
         for location in env.locations
         if location.label == "Classroom"
     )
