+++ tests/test_location_designer/test_combinations/test_split_nest.py
@@ -20,7 +20,7 @@ def test_1():
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-        n_actors = 2
+       # n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -34,7 +34,7 @@ def test_1():
     )
 
     assert len(env.actors) == 8
-    assert len(env.locations) == 6
+    assert len(env.locations) == 4
 
     for location in env.locations:
         if location.label == "School":
@@ -44,7 +44,7 @@ def test_1():
             assert counter[2] == 2
 
     assert not all(
-        location.actors[0].School == location.actors[1].School
+        location.actors[0].School == location.actors[-1].School
         for location in env.locations
         if location.label == "Classroom"
     )
@@ -53,7 +53,7 @@ def test_1():
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
