+++ tests/test_location_designer/test_combinations/test_overcrowding_nest.py
@@ -57,7 +57,7 @@ def test_2():
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -71,7 +71,7 @@ def test_2():
     )
 
     assert len(env.actors) == 8
-    assert len(env.locations) == 6
+    assert len(env.locations) == 4
 
     for location in env.locations:
         if location.label == "School":
@@ -81,20 +81,20 @@ def test_2():
             assert counter[2] == 2
 
     assert not all(
-        location.actors[0].School == location.actors[1].School
+        location.actors[0].School == location.actors[-1].School
         for location in env.locations
         if location.label == "Classroom"
     )
 
     inspector = p2n.NetworkInspector(env)
     inspector.plot_bipartite_network()
-    inspector.plot_actor_network(actor_attrs=df.columns, actor_color="id")
+    inspector.plot_actor_network(actor_attrs=df.columns, actor_color="_id")
 
     class School(p2n.LocationDesigner):
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -113,7 +113,7 @@ def test_2():
     assert len(env.actors) == 8
     assert len(env.locations) == 6
     assert all(
-        location.actors[0].School == location.actors[1].School
+        location.actors[0].School == location.actors[-1].School
         for location in env.locations
         if location.label == "Classroom"
     )
@@ -152,7 +152,7 @@ def test_3():
         n_actors = 4
 
     class Group(p2n.LocationDesigner):
-        n_actors = 2
+        #n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -180,7 +180,7 @@ def test_3():
         assert location.actors[0].group == location.actors[1].group
 
     # not all members of the same group are also in the same city (which is not desired)
-    assert not all(location.actors[0].City == location.actors[1].City for location in env.locations)
+    assert not all(location.actors[0].City == location.actors[-1].City for location in env.locations)
 
     class GroupNestedInCity(Group):
         def nest(self):
@@ -203,4 +203,4 @@ def test_3():
         assert location.actors[0].group == location.actors[1].group
 
     # all members of a group are in the same city
-    assert all(location.actors[0].City == location.actors[1].City for location in env.locations)
+    assert all(location.actors[0].City == location.actors[-1].City for location in env.locations)
