### 2025-05-23 03b7622
auto-commit: fixes by ruff
