+++ tests/test_location_designer/test_combinations/test_split_nest.py
@@ -20,7 +20,7 @@ def test_1():
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-       # n_actors = 2
+        # n_actors = 2
 
         def split(self, actor):
             return actor.group
@@ -53,7 +53,7 @@ def test_1():
         n_actors = 4
 
     class Classroom(p2n.LocationDesigner):
-        #n_actors = 2
+        # n_actors = 2
 
         def split(self, actor):
             return actor.group
