+++ tests/test_weighting/test_weighting_during_sim.py
@@ -8,18 +8,18 @@ def test_weighting_during_sim1():
         def setup(self):
             self.env = p2n.Environment()
             self.location = p2n.Location()
-            self.agent = p2n.Agent()
-            self.env.add_agent(self.agent)
+            self.actor = p2n.Actor()
+            self.env.add_actor(self.actor)
             self.env.add_location(self.location)
-            self.location.add_agent(agent=self.agent, weight=None)
+            self.location.add_actor(actor=self.actor, weight=None)
 
         def step(self):
-            self.location.set_weight(agent=self.agent, weight=None)
+            self.location.set_weight(actor=self.actor, weight=None)
 
         def update(self):
-            assert self.location.get_weight(agent=self.agent) == 1
-            assert self.agent.get_location_weight(location=self.location) == 1
-            assert self.env.get_weight(location=self.location, agent=self.agent) == 1
+            assert self.location.get_weight(actor=self.actor) == 1
+            assert self.actor.get_location_weight(location=self.location) == 1
+            assert self.env.get_weight(location=self.location, actor=self.actor) == 1
 
     model = WeightedModel()
     model.run(steps=10)
@@ -30,37 +30,43 @@ def test_weighting_during_sim2():
         def setup(self):
             self.env = p2n.Environment()
             self.location = p2n.Location()
-            self.agent = p2n.Agent()
-            self.env.add_agent(self.agent)
+            self.actor = p2n.Actor()
+            self.env.add_actor(self.actor)
             self.env.add_location(self.location)
-            self.location.add_agent(agent=self.agent, weight=self.t)
+            self.location.add_actor(actor=self.actor, weight=self.t)
 
         def step(self):
-            self.location.set_weight(agent=self.agent, weight=self.t)
+            self.location.set_weight(actor=self.actor, weight=self.t)
 
         def update(self):
-            assert self.location.get_weight(agent=self.agent) == self.t
-            assert self.agent.get_location_weight(location=self.location) == self.t
-            assert self.env.get_weight(location=self.location, agent=self.agent) == self.t
+            assert self.location.get_weight(actor=self.actor) == self.t
+            assert self.actor.get_location_weight(location=self.location) == self.t
+            assert self.env.get_weight(location=self.location, actor=self.actor) == self.t
 
     model = WeightedModel()
     model.run(steps=10)
 
 
 def test_weighting_during_sim3():
-    class WeightedModel(p2n.Model):
+    class WeightedModel(ap.Model):
         def setup(self):
-            self.location = p2n.Location(model=self)
-            self.agent = p2n.Agent(model=self)
-            self.location.add_agent(agent=self.agent, weight=10)
+            self.env = p2n.Environment(model=self)
+            
+            self.location = p2n.Location()
+            self.env.add_location(self.location)
+            
+            self.actor = p2n.Actor()
+            self.env.add_actor(self.actor)
+
+            self.location.add_actor(actor=self.actor, weight=10)
 
         def step(self):
             pass
 
         def update(self):
-            assert self.location.get_weight(agent=self.agent) == 10
-            assert self.agent.get_location_weight(location=self.location) == 10
-            assert self.get_weight(location=self.location, agent=self.agent) == 10
+            assert self.location.get_weight(actor=self.actor) == 10
+            assert self.actor.get_location_weight(location=self.location) == 10
+            assert self.env.get_weight(location=self.location, actor=self.actor) == 10
 
     model = WeightedModel()
     model.run(steps=10)
@@ -68,22 +74,28 @@ def test_weighting_during_sim3():
 
 def test_weighting_during_sim4():
     class WeightedLocation(p2n.Location):
-        def weight(self, agent):
-            return self.model.t
[…]
