### 2025-05-23 1711735
auto-commit: fixes by ruff
