### 2025-05-27 c0dce1f
fix .n_actors = 0 bug
