### 2025-05-27 5b02cac
Remove old python versions from readme
