+++ README.md
@@ -6,7 +6,8 @@
 Pop2net is a Python package that combines many steps related to network generation and management for agent-based modeling using a bipartite approach: 
 In Pop2net, relations are represented as a graph of agents and locations. Locations act as a contact layer, representing specific contexts through which agents connect. 
 This makes it easy to generate and manage custom network structures in an agent-based model.
-Pop2net extends the existing agent-based modeling framework [AgentPy](https://github.com/jofmi/agentpy), while also providing functionality for generating networks compatible with other modeling frameworks.
+
+Pop2net can be used within custom modeling frameworks and also integrates seamlessly with established agent-based modeling frameworks [AgentPy](https://github.com/jofmi/agentpy) and [Mesa](https://github.com/projectmesa/mesa).
 
 
 ## Installation
