### 2025-06-02 db6f205
fixed inspector.location_crosstab
