### 2025-06-02 c9a7422
fixed inspector.location_information
