### 2025-06-02 1be59bb
auto-commit: fixes by ruff
