### 2025-06-04 43bebcb
add the case where a model is passed to the env but not framework is used
