### 2025-06-10 7b794f0
auto-commit: fixes by ruff
