### 2025-06-11 9683fdd
add max to acknowledgements
