+++ paper/review.md
@@ -1,0 +1,28 @@
+reviewer 1:
+
+- [ ] Add the workflow visualization diagram - The about page in the documentation includes a helpful workflow diagram showing the relationship between micro-level data, agent classes, and location instances. Including this visualization (or a simplified version) would give readers a clearer understanding of Pop2net's approach.
+
+- [ ] Incorporate practical code examples - Your paper would benefit from showing how the concepts translate into straightforward code: ... These examples would demonstrate the simplicity of the API and make the bipartite approach more concrete. Here's an example how we did this in our Mesa paper (we might have done a little much).
+
+- [ ] Highlight location types as organizational tools - The paper mentions location types but could better explain how they effectively organize different relationship contexts. An example or visual might also help here: ... This pattern enables modelers to implement context-specific behaviors, a notable advantage of your approach.
+
+- [ ] Clarify distinction from population synthesizers - The documentation states that "Pop2net is not a population synthesizer." Adding this distinction to your paper would help readers better understand your tool's purpose and scope.
+
+- [ ] Expand on NetworkInspector functionality - The paper briefly mentions "inspection tools," but could benefit from highlighting the visualization capabilities (inspector.plot_bipartite_network() / inspector.plot_agent_network()) and maybe include a few figures as examples.
+
+- [ ] Better explain relationship with AgentPy - The documentation describes Pop2net as "a plugin for AgentPy" while the paper mentions "extending" it. A clearer explanation of how a ABM / AgentPy developer might use Pop2net would be useful for your target audience.
+
+- [ ] Extending that, maybe also discuss why AgentPy specifically was extended in the first place, and if it could also work with other Python ABM libraries (like Mesa for example).
+
+- [ ] Showcase network management features - The paper could demonstrate the intuitive API for common operations: ... These methods significantly reduce the code typically needed for network manipulation.
+
+- [ ] Illustrate bipartite-to-agent network projection - The documentation shows how Pop2net handles the projection from bipartite to agent networks. This feature deserves mention as it simplifies network analysis.
+
+
+reviewer 2:
+
+- [ ] Please provide a short explanation in the beginning of what a bipartite network is. I think for a newcomer it would not immediately be clear. Maybe a graphic or example could be helpful.
+
+- [ ] In the statement of need you say that other packages lack generalization while also being too abstract. These two critiques seem to contradict each other. Also, it does not seem that pop2net is about generalization, as it focuses on only one type of network. Maybe the statement of need could be improved by instead focusing on the advantages of pop2net's modelling approach (i.e. the three core features) in comparison to e.g. just using networkx graphs.
+
+- [ ] "while also providing functionality for generating networks compatible with other modeling frameworks" - this sounds cool but i couldn't find an example how and with which other framework it can be used. As @EwoutH mentioned, it would be cool of course if you could do something like Model.agentpy() and Model.mesa() to use pop2net with different base packages, but that is probably a nice-to-have for the future. But I think you need to illustrate what is meant by this sentence.
