### 2025-06-11 8cb1a5f
Better explain bipartite networks & remove ref to AgentPy in intro
