### 2025-06-12 #7897 @mariuzka
Hello @danielskatz, we are in the process of revising the paper. It should be completed within the next two weeks. Inspired by the comments, we have first implemented some major code changes over the last weeks and updated the documentation accordingly.
