### 2025-06-12 6bdee99
added tests for inspector.location_crosstab and location_information
