### 2025-06-12 e0f80b2
auto-commit: fixes by ruff
