+++ tests/test_inspector/test_location_information.py
@@ -1,15 +1,15 @@
 import pandas as pd
-from pop2net.inspector import NetworkInspector
+
 from pop2net.actor import Actor
-from pop2net.location import Location
 from pop2net.environment import Environment
+from pop2net.inspector import NetworkInspector
+from pop2net.location import Location
 
 
 def test_location_information_df():
-    
     def setup_env():
         env = Environment()
-        a1 = Actor() 
+        a1 = Actor()
         a1.status = "pupil"
         a2 = Actor()
         a2.status = "teacher"
@@ -19,13 +19,11 @@ def test_location_information_df():
         env.add_location(loc)
         [env.add_actor_to_location(loc, actor) for actor in env.actors]
         return env
-    
+
     env = setup_env()
     inspector = NetworkInspector(env)
     df = inspector.location_information(
-        location_labels="TestLoc",
-        actor_attributes=["status", "id_p2n"],
-        output_format="df"
+        location_labels="TestLoc", actor_attributes=["status", "id_p2n"], output_format="df"
     )
     assert isinstance(df, pd.DataFrame)
     # expect 2 rows, one per actor
@@ -38,11 +36,11 @@ def test_location_information_df():
     # statuses present
     assert set(df["status"]) == {"teacher", "pupil"}
 
-def test_location_information_table(capsys):
 
+def test_location_information_table(capsys):
     def setup_env():
         env = Environment()
-        a1 = Actor() 
+        a1 = Actor()
         a1.status = "pupil"
         a2 = Actor()
         a2.status = "teacher"
@@ -56,9 +54,7 @@ def test_location_information_table(capsys):
     env = setup_env()
     inspector = NetworkInspector(env)
     result = inspector.location_information(
-        location_labels="TestLoc",
-        actor_attributes="status",
-        output_format="table"
+        location_labels="TestLoc", actor_attributes="status", output_format="table"
     )
     # table mode returns None and prints header + table
     assert result is None
