+++ paper/paper.md
@@ -42,34 +42,40 @@ Bipartite networks consist of two distinct types of entities, where edges are on
 In Pop2net, relationships are represented as bipartite networks connecting *actors* and *locations*, with direct links existing only between actors and locations.
 However, when two actors are linked to the same location, they are considered indirectly connected through that shared location.
 In this way, locations serve as a contact layer between actors, representing the places where interactions occur or the contexts that facilitate actor connections.
-The bipartite approach to relations makes it easy to generate and manage custom network structures in agent-based models.
+The aim of Pop2net's bipartite approach to relations is to simplify the generation and management of network structures in ABM.
 
 
 # Statement of Need
 
-Currently, there are only a few tools designed to create and manage networks in agent-based models.
-General-purpose ABM frameworks, such as NetLogo [@wilensky_netlogo_1999], Mesa [@kazil_utilizing_2020], AgentPy [@foramitti_agentpy_2021], and Melodie [@yu_melodie_2023], offer basic data types for representing networks but lack advanced tools for creating custom network structures and provide only limited support for bipartite graphs out of the box.
-Existing network generators designed specifically for ABM, such as SynthPops [@mistry_synthpops_2021], often lack generalization, being domain-specific and limited in scope.
-As a result, most agent-based models rely on highly abstract network models [@amblard_which_2015].
-Custom modifications to tailor such network models to specific research designs are often complex and inflexible.
+Besides NetLogo [@wilensky_netlogo_1999], Python has become one of the primary programming languages for building agent-based models.
+In the Python ecosystem, NetworkX [@hagberg_exploring_2008] is the dominant library for working with networks.
+It provides extensive control over network structures and a wide range of ready-to-use network models.
+However, NetworkX was not originally designed with ABM in mind, which can make managing networks or creating custom network structures within ABM frameworks cumbersome and inflexible.
+While the built-in network models are powerful, they are often too abstract to directly support the specific needs of ABM applications.
 
-Pop2net aims to address this gap in the toolkit of agent-based modelers by offering a framework for creating and managing custom network structures, with the flexibility to integrate empirical data and traditional network models.
-We identify three core features of Pop2net that are currently lacking in existing software for ABM:
+ABM frameworks in Python—such as Mesa [@kazil_utilizing_2020], AgentPy [@foramitti_agentpy_2021], and Melodie [@yu_melodie_2023]—offer improved data structures for representing and manipulating networks.
+However, they still lack advanced tools for generating custom network topologies and provide only limited out-of-the-box support for bipartite graphs.
 
-1. **A bipartite approach to networks.**
-    Pop2net implements a bipartite approach to networks, which simplifies the creation and management of relations between agents in simulation models.
-    The approach of creating a network based on contact-layers is not entirely new, but it is already common in epidemiological agent-based models (e.g., @kerr_covasim_2020, @vermeulen_social_2021, @kaffai_modeling_2021).
-    However, Pop2net is the first software package to implement this approach in a general package for network generation and management in ABM.
+Pop2net aims to address this gap in the toolkit of agent-based modelers by utilizing bipartite networks for organizing relations in ABM, with the flexibility to integrate empirical data and traditional network models:
 
-2. **Scalable and modular network generation.**
+1. **Bringing bipartite networks to agent-based modeling in Python**
+    Pop2net implements a bipartite approach to networks in ABM, which can simplify their creation and management by organizing relations through locations.
+    Especially in models that simulate humans interacting in specific places or connected through shared contexts, organizing relationships through locations can be far more intuitive and easier to manage.
+    While this approach is already common in epidemiological ABM (e.g., @kerr_covasim_2020, @vermeulen_social_2021, @kaffai_modeling_2021), 
+    Pop2net is the first software package to generalize the bipartite approach for ABM. 
[…]
