### 2025-06-12 985dc4e
add ref to new mesa paper
