+++ paper/paper.bib
@@ -1,180 +1,4 @@
 
-@article{vermeulen_social_2021,
-	title = {Social {Network} {Metric}-{Based} {Interventions}? {Experiments} with an {Agent}-{Based} {Model} of the {COVID}-19 {Pandemic} in a {Metropolitan} {Region}},
-	volume = {24},
-	issn = {1460-7425},
-	shorttitle = {Social {Network} {Metric}-{Based} {Interventions}?},
-	doi = {10.18564/jasss.4571},
-	number = {3},
-	journal = {Journal of Artificial Societies and Social Simulation},
-	author = {Vermeulen, Ben and Müller, Matthias and Pyka, Andreas},
-	year = {2021},
-	keywords = {Agent-Based Model, Coronavirus, COVID-19, Epidemic, Policy Laboratory},
-	pages = {6},
-	file = {Social Network Metric-Based Interventions?:C\:\\Users\\ac135963\\Zotero\\storage\\YNREPJ6M\\6.html:text/html;Vermeulen et al_2021_Social Network Metric-Based Interventions.pdf:C\:\\Users\\ac135963\\Zotero\\storage\\5VUBUQ7A\\Vermeulen et al_2021_Social Network Metric-Based Interventions.pdf:application/pdf},
-}
-
-@article{dorn_wenn_nodate,
-	title = {Wenn {Menschen} (keine) {Menschen} treffen: {Simulation} der {Auswirkungen} von {Politikmaßnahmen} zur {Eindämmung} der zweiten {Covid}-19-{Welle}},
-	language = {de},
-	author = {Dorn, Florian and Gabler, Janos},
-	pages = {9},
-}
-
-@article{kaffai_modeling_2021,
-	title = {Modeling non-pharmaceutical interventions in the {COVID}-19 pandemic with survey-based simulations},
-	volume = {16},
-	issn = {1932-6203},
-	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0259108},
-	doi = {10.1371/journal.pone.0259108},
-	abstract = {Governments around the globe use non-pharmaceutical interventions (NPIs) to curb the spread of coronavirus disease 2019 (COVID-19) cases. Making decisions under uncertainty, they all face the same temporal paradox: estimating the impact of NPIs before they have been implemented. Due to the limited variance of empirical cases, researchers could so far not disentangle effects of individual NPIs or their impact on different demographic groups. In this paper, we utilize large-scale agent-based simulations in combination with Susceptible-Exposed-Infectious-Recovered (SEIR) models to investigate the spread of COVID-19 for some of the most affected federal states in Germany. In contrast to other studies, we sample agents from a representative survey. Including more realistic demographic attributes that influence agents’ behavior yields accurate predictions of COVID-19 transmissions and allows us to investigate counterfactual what-if scenarios. Results show that quarantining infected people and exploiting industry-specific home office capacities are the most effective NPIs. Disentangling education-related NPIs reveals that each considered institution (kindergarten, school, university) has rather small effects on its own, yet, that combined openings would result in large increases in COVID-19 cases. Representative survey-characteristics of agents also allow us to estimate NPIs’ effects on different age groups. For instance, re-opening schools would cause comparatively few infections among the risk-group of people older than 60 years.},
-	language = {en},
-	number = {10},
-	urldate = {2022-03-22},
-	journal = {PLOS ONE},
-	author = {Kaffai, Marius and Heiberger, Raphael H.},
-	month = oct,
-	year = {2021},
-	note = {Publisher: Public Library of Science},
[…]
