+++ _sources/Introduction/network_export.ipynb.txt
@@ -6,50 +6,101 @@
    "source": [
     "# Network model export\n",
     "\n",
-    "A network generated with Pop2net can be exported as a networkx graph object using `model.export_bipartite_network()` and `model.export_agent_network()`.\n",
+    "A network generated with Pop2net can be exported as a networkx graph object using `model.export_bipartite_network()` and `model.export__network()`.\n",
     "\n",
     "First, let us create a example network:"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 6,
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "LocationList (50 objects)"
+       "[<pop2net.creator.Location object at 0x7cf82757e450>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275811f0>,\n",
+       " <pop2net.creator.Location object at 0x7cf8276458b0>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275806e0>,\n",
+       " <pop2net.creator.Location object at 0x7cf82ad1f950>,\n",
+       " <pop2net.creator.Location object at 0x7cf82757ef30>,\n",
+       " <pop2net.creator.Location object at 0x7cf827580fe0>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275808c0>,\n",
+       " <pop2net.creator.Location object at 0x7cf827580920>,\n",
+       " <pop2net.creator.Location object at 0x7cf827580980>,\n",
+       " <pop2net.creator.Location object at 0x7cf82766f8c0>,\n",
+       " <pop2net.creator.Location object at 0x7cf82766f5f0>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275814f0>,\n",
+       " <pop2net.creator.Location object at 0x7cf827581610>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275815b0>,\n",
+       " <pop2net.creator.Location object at 0x7cf827581760>,\n",
+       " <pop2net.creator.Location object at 0x7cf827581460>,\n",
+       " <pop2net.creator.Location object at 0x7cf827581b80>,\n",
+       " <pop2net.creator.Location object at 0x7cf827714140>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275800b0>,\n",
+       " <pop2net.creator.Location object at 0x7cf827581fa0>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275807d0>,\n",
+       " <pop2net.creator.Location object at 0x7cf827582000>,\n",
+       " <pop2net.creator.Location object at 0x7cf827582060>,\n",
+       " <pop2net.creator.Location object at 0x7cf8275820c0>,\n",
+       " <pop2net.creator.Location object at 0x7cf82757e570>,\n",
+       " <pop2net.creator.Location object at 0x7cf827580c20>,\n",
[…]
