+++ _sources/Introduction/framework_integration.ipynb.txt
@@ -7,9 +7,7 @@
    "source": [
     "# Integration into ABM frameworks\n",
     "\n",
-    "Pop2net provides seamless integration with two widely used agent-based modeling frameworks: [Mesa](https://mesa.readthedocs.io/latest/) and [AgentPy](https://agentpy.readthedocs.io/en/latest/).\n",
-    "\n",
-    "To use Pop2net with either framework, you only need to follow two steps:\n",
+    "Pop2net provides seamless integration with two widely used agent-based modeling frameworks: [Mesa](https://mesa.readthedocs.io/latest/) and [AgentPy](https://agentpy.readthedocs.io/en/latest/). To use Pop2net with either framework, you only need to follow two steps:\n",
     "\n",
     "First, when defining an Actor or Location class, it must inherit from both the corresponding Pop2net class and the framework’s class — in that order. For example:\n",
     "\n",
@@ -48,7 +46,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
+   "execution_count": 162,
    "id": "a5fe9017",
    "metadata": {},
    "outputs": [],
@@ -60,7 +58,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 163,
    "id": "fad5438b",
    "metadata": {},
    "outputs": [],
@@ -80,7 +78,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 164,
    "id": "e1229c4a",
    "metadata": {},
    "outputs": [],
@@ -101,7 +99,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 165,
    "id": "1fdfe213",
    "metadata": {},
    "outputs": [],
@@ -136,14 +134,14 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 166,
    "id": "1692ccab",
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "[<__main__.Actor at 0x126748ec0>, <__main__.Location at 0x126749010>]"
+       "[<__main__.Actor at 0x133bca900>, <__main__.Location at 0x133bcaa50>]"
       ]
      },
      "execution_count": null,
@@ -169,14 +167,14 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 167,
    "id": "d7868bda",
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "<mesa.agent.AgentSet at 0x115cc2d70>"
+       "<mesa.agent.AgentSet at 0x133d573b0>"
       ]
      },
      "execution_count": null,
@@ -190,14 +188,14 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 168,
    "id": "5070bb80",
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "[<__main__.Actor at 0x126748ec0>]"
+       "[<__main__.Actor at 0x133bca900>]"
       ]
      },
      "execution_count": null,
@@ -211,14 +209,14 @@
   },
   {
[…]
