+++ src/pop2net/inspector.py
@@ -344,7 +344,11 @@ class NetworkInspector:
         if return_df:
             return df
 
-    def network_measures(self, actor_attrs=None) -> list[dict]:
+    def network_measures(
+            self,
+            actor_attrs:list[str] =None,
+            weighted: bool  = True
+        ) -> list[dict]:
         """Calculates common network measures for the actor-level network graph.
 
         If the created network consist of independent groups of nodes
@@ -352,6 +356,7 @@ class NetworkInspector:
 
         Args:
             actor_attrs: A list of actor attributes
+            weighted: Should the network be considered weighted or unweighted
 
         Return:
             list of dictionaries of the common network measure results
@@ -361,6 +366,7 @@ class NetworkInspector:
         # make distinction between multiple independent networks and one network
 
         def get_network_measures(nx_graph) -> dict:
+
             result_dict = {}
             result_dict["n_nodes"] = nx.number_of_nodes(nx_graph)
             result_dict["diameter"] = nx.diameter(nx_graph, weight="weight")
@@ -375,6 +381,21 @@ class NetworkInspector:
                 weight="weight",
             )
             return result_dict
+        
+        def get_network_measures_no_weight(nx_graph) -> dict:
+            
+            result_dict = {}
+            result_dict["n_nodes"] = nx.number_of_nodes(nx_graph)
+            result_dict["diameter"] = nx.diameter(nx_graph, weight = None)
+            result_dict["density"] = nx.density(nx_graph)
+            result_dict["transitivity"] = nx.transitivity(nx_graph)
+            result_dict["avg_clustering"] = nx.average_clustering(
+                nx_graph,
+            )
+            result_dict["avg_path_length"] = nx.average_shortest_path_length(
+                nx_graph,
+            )
+            return result_dict
 
         network_components = list(nx.connected_components(nx_graph))
         network_components = sorted(network_components, key=len, reverse=True)
@@ -384,11 +405,14 @@ class NetworkInspector:
         for component in network_components:
             try:
                 nx_subgraph = nx_graph.subgraph(component)
+                
             except nx.NetworkXError:
                 print("Cant make graph out of component")
                 break
-
-            result_list.append(get_network_measures(nx_subgraph))
+            if weighted:
+                result_list.append(get_network_measures(nx_subgraph))
+            else:
+                result_list.append(get_network_measures_no_weight(nx_subgraph))
         return result_list
 
     def location_crosstab(
