### 2025-06-16 79d878a
auto-commit: fixes by ruff
