### 2025-06-17 fefceb9
improve flow chart and insert it into the paper
