### 2025-06-27 02f417f
remove basic code example
