### 2025-06-28 8807663
some fixes on paper
