+++ paper/paper.md
@@ -19,7 +19,7 @@ authors:
     orcid: 0000-0002-4977-2947
     equal-contrib: true
     affiliation: 1
-  
+
 affiliations:
  - name: University of Stuttgart, Germany
    index: 1
@@ -32,26 +32,25 @@ bibliography: paper.bib
 
 # Summary
 
-Agent-based modeling (ABM) is a scientific method used in fields such as social science, biology, and ecology to simulate the interactions of autonomous agents and study the resulting emergent phenomena.
-The relationships between agents that structure the simulated interactions are often represented by network graphs.
-Since empirical data on networks is rare, most of agent-based models rely on artificially generated networks [@amblard_which_2015].
-Consequently, generating a valid network structure at the beginning of a simulation, as well as accessing and modifying it during the simulation, are critical steps that must be managed in almost any agent-based model.
+Agent-based modeling (ABM) is a scientific method used in fields such as social science, biology, and ecology to simulate interactions of autonomous agents and study the resulting emergent phenomena.
+The relationships between agents, which structure the simulated interactions, are often represented by network graphs.
+Since empirical data on networks is rare, most agent-based models rely on artificially generated networks [@amblard_which_2015].
+Consequently, generating a valid network structure at the beginning of a simulation is critical.
+Addtionally, accessing and modifying the network during the simulation are steps that must be managed in almost any agent-based model.
 
 Pop2net is a Python package that combines many steps related to network generation and management for ABM using a bipartite approach.
-Bipartite networks consist of two distinct types of entities, where edges are only formed between entities of different types.
+Bipartite networks consist of two distinct types of entities where edges are only formed between entities of different types.
 In Pop2net, relationships are represented as bipartite networks connecting *actors* and *locations*, with direct links existing only between actors and locations.
-However, when two actors are linked to the same location, they are considered indirectly connected through that shared location.
-In this way, locations serve as a contact layer between actors, representing the places where interactions occur or the contexts that facilitate actor connections.
+However, when two actors are linked to the same location, they are considered connected through that shared location.
+In this way, locations serve as a contact layer between actors, representing places where interactions occur or contexts that facilitate actor connections.
 The aim of Pop2net's bipartite approach to relations is to simplify the generation and management of network structures in ABM.
 
 
 # Statement of Need
 
-<!-- TODO: Den ersten Absatz kürzen -->
-Besides NetLogo [@wilensky_netlogo_1999], Python has become one of the primary programming languages for building agent-based models.
-In the Python ecosystem, NetworkX [@hagberg_exploring_2008] is the dominant library for working with networks.
+Building upon NetworkX [@hagberg_exploring_2008], we provide an extensive tool designed to simplify and streamline the generation of (empirically calibrated) bipartite networks for further use with common ABM frameworks.
 It provides extensive control over network structures and a wide range of ready-to-use network models.
-However, NetworkX was not originally designed with ABM in mind, which can make managing networks or creating custom network structures within ABM frameworks cumbersome and inflexible.
+NetworkX was not originally designed with ABM in mind, which can make managing networks or creating custom network structures within ABM frameworks cumbersome and inflexible.
 While the built-in network models are powerful, they are often too abstract to directly support the specific needs of ABM applications.
 
 ABM frameworks in Python—such as Mesa [@kazil_utilizing_2020, @hoeven_mesa_2025], AgentPy [@foramitti_agentpy_2021], and Melodie [@yu_melodie_2023]—offer improved data structures for representing and manipulating networks.
@@ -62,19 +61,19 @@ Pop2net aims to address this gap in the toolkit of agent-based modelers by utili
 1. **Bringing bipartite networks to agent-based modeling in Python**
[…]
