### 2025-06-29 fac9ba1
update code example
