### 2025-06-29 325165f
display network plots side by side
