+++ paper/paper.md
@@ -59,19 +59,19 @@ Pop2net aims to address this gap in the toolkit of agent-based modelers.
 Building upon NetworkX, we provide an extensive tool designed to simplify and streamline the generation of (empirically calibrated) bipartite networks for further use with common ABM frameworks.
 While Pop2net is implemented in Python, to our knowledge, no comparable software exists in other programming environments that offers the following features:
 
-1. **Bringing bipartite networks to agent-based modeling**
+1. **Bringing bipartite networks to agent-based modeling.**
     Pop2net implements a bipartite approach to networks in ABM, which can simplify their creation and management by organizing relations through locations.
     Especially in models that simulate humans interacting in specific places or connected through shared contexts, organizing relationships through locations can be far more intuitive and easier to manage.
     While this approach is already common in epidemiological ABM (e.g., @kerr_covasim_2020, @vermeulen_social_2021, @kaffai_modeling_2021),
     Pop2net is the first software package to generalize the bipartite approach for ABM.
     In addition, networks generated by Pop2net—consisting of actors and locations—can be seamlessly integrated into Mesa and AgentPy.
 
-2. **Generating scalable and modular networks**
+2. **Generating scalable and modular networks.**
     Based on the bipartite approach, Pop2net offers an innovative way of creating networks programmatically and modularly.
     Users can quickly create complex networks based on different location definitions that are translated into network structures.
     Each location definition contains a set of rules for how agents connect, and can integrate empirical data as well as classic network models.
 
-3. **Integrating micro-level data**
+3. **Integrating micro-level data.**
     Pop2net is designed to integrate empirical micro-level data, e.g., from surveys, into the population and network generation processes.
     The empirical micro-level data can be used to generate a population of actors and then refer to the empirical attributes when determining the rules of how actors connect via location definitions.
     This enables users, especially those from the social sciences, to ground network generation in empirical data, even in the absence of detailed network data.
@@ -81,7 +81,7 @@ While Pop2net is implemented in Python, to our knowledge, no comparable software
 
 Pop2net's components can be categorized into three sectors.
 
-1. **Environment, Actor and Location.**
+1. **Basic elements for bipartite networks in ABM.**
     The first sector contains the three object classes—Environment, Actor, and Location—that must always be used to facilitate the bipartite network structure of Pop2net.
     The Environment class is Pop2net's main object that holds all entities and manages their relations.
     Central attributes of the Environment class are:
@@ -96,7 +96,7 @@ Pop2net's components can be categorized into three sectors.
     Both actors and locations provide several methods which help managing relations during simulations—for example, to retrieve all associated locations or actors, find neighboring actors within specific location types, or connect actors to one another through certain locations.
 
 
-2. **Creator and LocationDesigner.**
+2. **Network generators.**
     The classes Creator and LocationDesigner are the tools in Pop2net that enable the user to generate custom (bipartite) networks in a modular and scalable way.
     By defining location types using the LocationDesigner, users can, for instance, quickly specify which agent should be connected to a certain location type, how many location instances should be created, whether locations are nested within other locations, or how strongly the connection between an actor and a location is weighted.
     Based on these definitions, the Creator class generates a bipartite network of actors and locations.
@@ -106,7 +106,7 @@ Pop2net's components can be categorized into three sectors.
     Those empirically created attributes support the creation of realistic network structures by including them in the location definitions, for instance, the age or the household of the actors.
     Location types can also incorporate network graphs based on empirical data or generated by NetworkX.
 
-3. **NetworkInspector.**
+3. **Inspection tools.**
[…]
