### 2025-06-30 0bc589a
make description of code example shorter
