+++ paper/paper.md
@@ -43,7 +43,7 @@ Bipartite networks consist of two distinct types of entities where edges are onl
 In Pop2net, relationships are represented as bipartite networks connecting *actors* and *locations*.
 When two actors are linked to the same location, they are considered connected through that shared location.
 In this way, locations serve as a contact layer between actors, representing places where interactions occur or contexts that facilitate actor connections.
-The aim of Pop2net's bipartite approach to relations is to simplify the generation and management of network structures in ABM.
+The aim of Pop2net's bipartite approach to relations is to simplify the generation and management of realistic network structures in ABM.
 
 
 # Statement of Need
@@ -63,7 +63,7 @@ While Pop2net is implemented in Python, to our knowledge, no comparable software
     Especially in models that simulate humans interacting in specific places or connected through shared contexts, organizing relationships through locations can be far more intuitive and easier to manage.
     While this approach is already common in epidemiological ABM (e.g., @kerr_covasim_2020, @vermeulen_social_2021, @kaffai_modeling_2021),
     Pop2net is the first software package to generalize the bipartite approach for ABM.
-    In addition, networks generated by Pop2net—consisting of actors and locations—can be seamlessly integrated into Mesa and AgentPy.
+    Pop2net can serve as the foundation for custom simulation frameworks or be seamlessly integrated into the widely used ABM frameworks Mesa and AgentPy.
 
 2. **Generating scalable and modular networks.**
     Based on the bipartite approach, Pop2net offers an innovative way of creating networks programmatically and modularly.
@@ -82,24 +82,28 @@ Pop2net's components can be categorized into three sectors.
 
 1. **Basic elements for bipartite networks in ABM.**
     The first sector contains the three object classes—Environment, Actor, and Location—that must always be used to facilitate the bipartite network structure of Pop2net.
+    
     The Environment class is Pop2net's main object that holds all entities and manages their relations.
+    It is designed for use within simulation models.
     Central attributes of the Environment class are:
     * A NetworkX graph object that stores all actors, locations, and their relationships.
     * An actor list and a location list that provide convenient access to actor and location objects.
     * Various methods for managing tasks such as connecting actors and locations or exporting the network.
 
     Actors are the (inter)acting entities in the simulation—what is typically referred to as *agents* in ABM.
-    However, since Pop2net also treats locations as agents, especially when integrated with frameworks like Mesa or AgentPy, we found it clearer to refer to the primary acting entities as *actors*.
+    However, since Pop2net also treats locations as *agents*, especially when integrated with frameworks like Mesa or AgentPy, we found it clearer to refer to the primary acting entities as *actors*.
     Locations represent the places or contexts in which actors interact.
     In Pop2net, every connection between actors must be mediated by a location.
+    
     Both actors and locations provide several methods which help managing relations during simulations—for example, to retrieve all associated locations or actors, find neighboring actors within specific location types, or connect actors to one another through certain locations.
+    Using multiple inheritance, actors and locations can also be integrated as agent objects into Mesa and AgentPy.
 
 
 2. **Network generators.**
     The classes Creator and LocationDesigner are the tools in Pop2net that enable the user to generate custom (bipartite) networks in a modular and scalable way.
     By defining location types using the LocationDesigner, users can, for instance, quickly specify which agent should be connected to a certain location type, how many location instances should be created, whether locations are nested within other locations, or how strongly the connection between an actor and a location is weighted.
     Based on these definitions, the Creator class generates a bipartite network of actors and locations.
-    The generated networks and their components can serve as the basis for simulation models, be integrated with Mesa and AgentPy, or be exported to NetworkX in both bipartite and unipartite formats.
[…]
