### 2025-07-01 f09c81c
add a intro section on surveys
