+++ _sources/api/location.rst.txt
@@ -3,7 +3,7 @@
 Location
 ========
 
-The location class represents a base class that can be used to implement realistic location in which agents can meet and interact.
+The location class represents a base class that can be used to implement realistic location in which actors can meet and interact.
 
 .. autoclass:: Location
     :members:
