+++ _sources/api/actor.rst.txt
@@ -1,0 +1,10 @@
+.. currentmodule:: pop2net
+
+=====
+Actor
+=====
+
+This is an overview over the `Actor` Class
+
+.. autoclass:: Actor
+    :members:
