### 2025-07-02 b1c9cc3
make returning nothing the default setting of creator.create()
