+++ src/pop2net/creator.py
@@ -944,6 +944,7 @@ class Creator:
         replace_sample_level_column: bool = True,
         clear: bool = False,
         delete_magic_actor_attributes: bool = True,
+        return_elements = False,
     ) -> tuple:
         """Creates actors and locations based on a given dataset.
 
@@ -1004,7 +1005,8 @@ class Creator:
             delete_magic_actor_attributes=delete_magic_actor_attributes,
         )
 
-        return actors, locations
+        if return_elements:
+            return actors, locations
 
     def get_df_actors(
         self,
