### 2025-07-02 a0d0675
auto-commit: fixes by ruff
