### 2025-07-02 a45b955
bump version to 1.0.2
