+++ _sources/Introduction/network_export.ipynb.txt
@@ -15,67 +15,7 @@
    "cell_type": "code",
    "execution_count": 1,
    "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/plain": [
-       "[<pop2net.creator.Location object at 0x13017da90>,\n",
-       " <pop2net.creator.Location object at 0x13017dbe0>,\n",
-       " <pop2net.creator.Location object at 0x13017dd30>,\n",
-       " <pop2net.creator.Location object at 0x13017de80>,\n",
-       " <pop2net.creator.Location object at 0x13017dfd0>,\n",
-       " <pop2net.creator.Location object at 0x13017d940>,\n",
-       " <pop2net.creator.Location object at 0x13017e270>,\n",
-       " <pop2net.creator.Location object at 0x13017e3c0>,\n",
-       " <pop2net.creator.Location object at 0x13017e510>,\n",
-       " <pop2net.creator.Location object at 0x13017e660>,\n",
-       " <pop2net.creator.Location object at 0x13017e120>,\n",
-       " <pop2net.creator.Location object at 0x13017e900>,\n",
-       " <pop2net.creator.Location object at 0x13017ea50>,\n",
-       " <pop2net.creator.Location object at 0x13017eba0>,\n",
-       " <pop2net.creator.Location object at 0x13017ecf0>,\n",
-       " <pop2net.creator.Location object at 0x13017e7b0>,\n",
-       " <pop2net.creator.Location object at 0x13017ef90>,\n",
-       " <pop2net.creator.Location object at 0x13017f0e0>,\n",
-       " <pop2net.creator.Location object at 0x13017f230>,\n",
-       " <pop2net.creator.Location object at 0x13017f380>,\n",
-       " <pop2net.creator.Location object at 0x13017ee40>,\n",
-       " <pop2net.creator.Location object at 0x13017f620>,\n",
-       " <pop2net.creator.Location object at 0x13017f770>,\n",
-       " <pop2net.creator.Location object at 0x13017f8c0>,\n",
-       " <pop2net.creator.Location object at 0x13017fa10>,\n",
-       " <pop2net.creator.Location object at 0x13017fb60>,\n",
-       " <pop2net.creator.Location object at 0x13017f4d0>,\n",
-       " <pop2net.creator.Location object at 0x13017fcb0>,\n",
-       " <pop2net.creator.Location object at 0x13017fe00>,\n",
-       " <pop2net.creator.Location object at 0x1301ec050>,\n",
-       " <pop2net.creator.Location object at 0x1301ec2f0>,\n",
-       " <pop2net.creator.Location object at 0x1301ec1a0>,\n",
-       " <pop2net.creator.Location object at 0x1301ec440>,\n",
-       " <pop2net.creator.Location object at 0x1301ec590>,\n",
-       " <pop2net.creator.Location object at 0x1301ec6e0>,\n",
[…]
