### 2025-08-05 79f91c0
Add custom list wrappers for actors and locations
Introduces ActorList and LocationList classes with custom string representations to improve clarity when inspecting actors and locations. The actors and locations properties now return these custom list types when no framework is used
