### 2025-08-05 2fbcfc0
auto-commit: fixes by ruff
