### 2025-09-01 ea0617d
put content of about page on the index page
