### 2025-09-03 b037be5
Add self-check to contact_weight method
Raises a ValueError if the provided actor is the same instance as self in the contact_weight method, preventing invalid comparisons.
