### 2025-09-03 5527bdb
auto-commit: fixes by ruff
