### 2025-09-05 afa6134
removed # %%
