### 2025-09-08 f4db798
added new print outs inside _to_framework
One unsolved problem are empty actor and location lists.

The type value of the first element in the object lists determines which __str__ and __repr__ will be set.
