### 2025-09-08 6fc62f1
Refactor custom list wrappers for environment objects
Replaces separate ActorList and LocationList classes with a unified ObjectList subclass for improved string representation of object lists. Simplifies the _to_framework method and removes redundant _to_custom_list logic, streamlining how actors and locations are returned when no framework is specified.
