### 2025-09-08 4aa3497
added test for new print and repr for env.actors/locations
