### 2025-09-09 6e33a2d
restructure new lists a little bit
