+++ tests/test_environment/test_print_out.py
@@ -10,10 +10,8 @@ def test_1():
     env.add_location(location1)
 
     location1.add_actors([actor1, actor2])
-    print(env.locations)
-    assert str(env.locations) == "ObjectList [1 Elements]"
-    assert repr(env.locations) == "ObjectList [1 Elements]"
-    assert str(env.actors) == "ObjectList [2 Elements]"
-    assert repr(env.actors) == "ObjectList [2 Elements]"
-
-    # TODO other framework prints? 
+    
+    assert str(env.locations) == "EntityList [1 entities]"
+    assert repr(env.locations) == "EntityList [1 entities]"
+    assert str(env.actors) == "EntityList [2 entities]"
+    assert repr(env.actors) == "EntityList [2 entities]"
