+++ tests/test_misc/test_entity_list.py
@@ -1,6 +1,6 @@
 import pop2net as p2n
 
-def test_1():
+def test_print_in_environment():
     env = p2n.Environment()
     actor1 = p2n.Actor()
     actor2 = p2n.Actor()
@@ -15,3 +15,9 @@ def test_1():
     assert repr(env.locations) == "EntityList [1 entities]"
     assert str(env.actors) == "EntityList [2 entities]"
     assert repr(env.actors) == "EntityList [2 entities]"
+
+
+def test_print_in_creator():
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
+    assert str(creator.create_actors(n=10)) == "EntityList [10 entities]"
