### 2025-09-09 2f74d8a
auto-commit: fixes by ruff
