+++ tests/test_misc/test_entity_list.py
@@ -1,5 +1,6 @@
 import pop2net as p2n
 
+
 def test_print_in_environment():
     env = p2n.Environment()
     actor1 = p2n.Actor()
@@ -10,7 +11,7 @@ def test_print_in_environment():
     env.add_location(location1)
 
     location1.add_actors([actor1, actor2])
-    
+
     assert str(env.locations) == "EntityList [1 entities]"
     assert repr(env.locations) == "EntityList [1 entities]"
     assert str(env.actors) == "EntityList [2 entities]"
@@ -25,4 +26,7 @@ def test_print_in_creator():
     creator = p2n.Creator(env=env)
 
     assert str(creator.create_actors(n=10)) == "EntityList [10 entities]"
-    assert str(creator.create_locations(location_designers=[TestLocationDesigner])) == "EntityList [1 entities]"
+    assert (
+        str(creator.create_locations(location_designers=[TestLocationDesigner]))
+        == "EntityList [1 entities]"
+    )
