### 2025-09-09 b4adceb
move check for same object
