+++ tests/test_environment/test_same_object_issue.py
@@ -1,0 +1,4 @@
+import pop2net as p2n
+
+def test_1():
+    env = p2n.Environment()
