+++ src/pop2net/actor.py
@@ -126,8 +126,6 @@ class Actor:
         Returns:
             A weight of the contact between the two actors.
         """
-        if actor is self:
-            raise ValueError("actor must be a different instance than self")
 
         weight = 0
         for location in self.shared_locations(actor=actor, location_labels=location_labels):
