+++ src/pop2net/environment.py
@@ -330,7 +330,6 @@ class Environment:
         )
 
     def _objects_between_objects(self, object1, object2) -> list:
-
         if object1 is object2:
             raise ValueError("Entity 1 and entity 2 are identical.")
 
