+++ tests/test_environment/test_same_object_issue.py
@@ -1,4 +1,5 @@
 import pop2net as p2n
 
+
 def test_1():
     env = p2n.Environment()
