+++ src/pop2net/actor.py
@@ -126,7 +126,6 @@ class Actor:
         Returns:
             A weight of the contact between the two actors.
         """
-
         weight = 0
         for location in self.shared_locations(actor=actor, location_labels=location_labels):
             weight += location.project_weights(actor1=self, actor2=actor)
