### 2025-09-10 45e323d
added two test cases
