### 2025-09-10 3ff8406
added exact error matching
