### 2025-09-10 972bd1e
added two more test, renamed tests for clarity
