### 2025-09-10 223ac91
auto-commit: fixes by ruff
