### 2025-09-11 bd624f8
tests added for shared_locations method
