+++ tests/test_environment/test_same_object_issue.py
@@ -47,3 +47,29 @@ def test_4_location_creator():
     with pytest.raises(ValueError) as exc:
         env.locations_between_actors(env.actors[0], env.actors[0])
     assert str(exc.value) == "Entity 1 and entity 2 are identical."
+
+def test_5_shared_loc():
+
+    env = p2n.Environment()
+    actor1 = p2n.Actor()
+    actor2 = p2n.Actor()
+    location1 = p2n.Location()
+    env.add_actors([actor1, actor2])
+    env.add_location(location1)
+    location1.add_actors([actor1, actor2])
+    with pytest.raises(ValueError) as exc:
+        env.actors[0].shared_locations(env.actors[0])
+    assert str(exc.value) == "Entity 1 and entity 2 are identical."
+    
+
+def test_6_shared_loc_creator():
+    class Location1(p2n.LocationDesigner):
+        pass
+
+    env = p2n.Environment()
+    creator = p2n.Creator(env=env)
+    creator.create_actors(n=2)
+    creator.create_locations(location_designers=[Location1])
+    with pytest.raises(ValueError) as exc:
+        env.actors[0].shared_locations(env.actors[0])
+    assert str(exc.value) == "Entity 1 and entity 2 are identical."
