### 2025-09-11 35c9d9a
auto-commit: fixes by ruff
