### 2025-09-12 b057eb8
bump version to 1.0.3
