### 2025-09-19 278a9a4
Improve EntityList string representation
EntityList __str__ now displays counts by type (actors, locations) instead of generic entity count. Updated related tests to match new output format for better clarity.
