+++ tests/test_misc/test_entity_list.py
@@ -12,10 +12,10 @@ def test_print_in_environment():
 
     location1.add_actors([actor1, actor2])
 
-    assert str(env.locations) == "EntityList [1 entities]"
-    assert repr(env.locations) == "EntityList [1 entities]"
-    assert str(env.actors) == "EntityList [2 entities]"
-    assert repr(env.actors) == "EntityList [2 entities]"
+    assert str(env.locations) == "EntityList [1 location]"
+    #assert repr(env.locations) == "EntityList [1 entities]"
+    assert str(env.actors) == "EntityList [2 actors]"
+    #assert repr(env.actors) == "EntityList [2 entities]"
 
 
 def test_print_in_creator():
@@ -25,8 +25,8 @@ def test_print_in_creator():
     env = p2n.Environment()
     creator = p2n.Creator(env=env)
 
-    assert str(creator.create_actors(n=10)) == "EntityList [10 entities]"
+    assert str(creator.create_actors(n=10)) == "EntityList [10 actors]"
     assert (
         str(creator.create_locations(location_designers=[TestLocationDesigner]))
-        == "EntityList [1 entities]"
+        == "EntityList [1 location]"
     )
