### 2025-09-19 6477d16
auto-commit: fixes by ruff
