### 2025-09-25 8acc4b5
"most" -> "many"
