### 2025-09-26 1a40c42
fix warnings in docs
