+++ docs/api/environment.rst
@@ -1,7 +1,7 @@
 .. currentmodule:: pop2net.environment
 
 Environment
-=====
+===========
 
 .. autoclass:: pop2net.Environment
     :members:
