+++ docs/api/location_designer.rst
@@ -1,7 +1,7 @@
 .. currentmodule:: pop2net
 
 LocationDesigner
-=============
+================
 
 The LocationDesigner class is an interface to design specific network structures using the Creator.
 
