### 2025-09-26 32124c4
extend "how to contribute" section in docs
