### 2025-09-26 1da571e
add section about documentation
