### 2025-09-26 #7897 @EwoutH
@luerhard Thanks for getting back!

It might be useful to link to your notes to developers directly in the Readme, but I ticked it off.

As for the Contribution and authorship, I would like @danielskatz and/or @hugoledoux to weigh in what the JOSS policy is on this. From what I could observe from the public git, Max Richter did significant feature work (but not as much as Marius Kaffai).
