+++ README.md
@@ -28,5 +28,5 @@ Detailed documentation of all features and a collection of example analyses can 
 ## How to contribute
 
 Contributions are welcome!  
-See the [this guide](https://mariuzka.github.io/pop2net/developers.html) 
+See [this guide](https://mariuzka.github.io/pop2net/developers.html) 
 for details on how to get started.
