### 2025-09-30 #7897 @mariuzka
Thanks @EwoutH for your feedback! I just added a link to the developers guide in the README.
