### 2025-09-30 e44147c
object type check changed to isinstance
