### 2025-10-02 d64b89f
added tests for combined objects and unknown objects
