### 2025-10-02 2060d10
auto-commit: fixes by ruff
