### 2025-10-03 646715a
test for entity list with loc/actor/unknown object combined
