### 2025-11-13 #7897 @hugoledoux
re the authorship, our rules are there: https://joss.readthedocs.io/en/latest/review_criteria.html#authorship

Usually we use them to *not* have authors who only provided the grant as we do not particularly like this practice that is very common in academia...

Here I would live it to the authors to decide. Personally I would add the SA if they did a lot of the work, I think it's nice and rewarding. But if the authors prefer an acknowlegement it's also fine.

@luerhard let us know
