### 2025-11-13 96e9970
Add Max as an author
