### 2025-12-05 734582f
Update README with JOSS badge
