+++ tests/test_misc/test_entity_list.py
@@ -13,9 +13,7 @@ def test_print_in_environment():
     location1.add_actors([actor1, actor2])
 
     assert str(env.locations) == "EntityList [1 location]"
-    # assert repr(env.locations) == "EntityList [1 entities]"
     assert str(env.actors) == "EntityList [2 actors]"
-    # assert repr(env.actors) == "EntityList [2 entities]"
 
     # mixed objects
     combined = p2n.EntityList([*env.locations, *env.actors])
