### 2025-12-15 7249334
move method and add one for locations
