### 2025-12-16 5a9dcec
auto-commit: fixes by ruff
