### 2025-12-16 08a4ad0
or-part removed
