### 2025-12-16 4a0f99e
auto-commit: fixes by ruff
