### 2025-12-19 239f12e
basic test for env.get_df_actors
