+++ tox.ini
@@ -4,7 +4,8 @@ envlist =
   py39,
   py310,
   py311,
-  py312-{numpy1,numpy2}
+  py312-{numpy1,numpy2},
+  py313-{numpy1,numpy2}
 
 [gh-actions]
 python =
@@ -12,6 +13,7 @@ python =
     3.10: py310
     3.11: py311
     3.12: py312
+    3.13: py313
 
 [testenv]
 deps =
