+++ tox.ini
@@ -5,7 +5,8 @@ envlist =
   py310,
   py311,
   py312-{numpy1,numpy2},
-  py313-{numpy1,numpy2}
+  py313-{numpy1,numpy2},
+  py314-{numpy1,numpy2}
 
 [gh-actions]
 python =
@@ -14,6 +15,7 @@ python =
     3.11: py311
     3.12: py312
     3.13: py313
+    3.14: py314
 
 [testenv]
 deps =
