+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
 isolated_build = True
+skip_missing_interpreters = false
 envlist =
   py39,
   py310,
