### 2025-12-30 717f112
remove python 3.14 & fix numpy versions
