+++ tox.ini
@@ -2,26 +2,20 @@
 isolated_build = True
 skip_missing_interpreters = false
 envlist =
-  py39,
-  py310,
   py311,
   py312-{numpy1,numpy2},
-  py313-{numpy1,numpy2},
-  py314-{numpy1,numpy2}
+  py313-numpy2,
 
 [gh-actions]
 python =
-    3.9: py39
-    3.10: py310
     3.11: py311
     3.12: py312
     3.13: py313
-    3.14: py314
 
 [testenv]
 deps =
-  numpy1: numpy==1.26.0
-  numpy2: numpy==2.0.1
+  numpy1: numpy==1.26.4
+  numpy2: numpy==2.3.4
 allowlist_externals = poetry
 commands_pre =
   poetry install --only dev
