### 2025-12-30 013f004
auto-commit: fixes by ruff
