### 2025-12-30 1beb7c4
print __str__() when __repr__() is called on entity lists
