### 2026-01-02 c6c03dc
test added for function env_get_df_locations
tests checks for correct content of the dataframe.
