### 2026-01-05 c34b1c3
update lockfile to py3.14 ready dependencies, update tox
