+++ tox.ini
@@ -5,12 +5,14 @@ envlist =
   py311,
   py312-{numpy1,numpy2},
   py313-numpy2,
+  py314-numpy2,
 
 [gh-actions]
 python =
     3.11: py311
     3.12: py312
     3.13: py313
+    3.14: py314
 
 [testenv]
 deps =
