+++ .github/workflows/tox.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.11', '3.12', '3.13']
+        python-version: ['3.11', '3.12', '3.13', '3.14.2']
     steps:
       #----------------------------------------------
       #       check-out repo and set-up python
