### 2026-01-09 #7897 @danielskatz
Congratulations to @mariuzka (Marius Kaffai) and co-authors on your JOSS publication!!

And thanks to @EwoutH and @jofmi for reviewing, and to @hugoledoux for editing.
JOSS completely depends on volunteers and we couldn't do this without you.
